% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf-viz.R
\name{add.perc}
\alias{add.perc}
\title{Add a percentage to counts}
\usage{
add.perc(x, ...)
}
\arguments{
\item{mafviz}{barplot An of the mafviz plot that follow barplot visualization}
}
\value{
mafviz Barplot The same barplot is now returned with percentages
}
\description{
Add a percentage to counts
}
\examples{
ggvarclass(mut) + add.perc()

}
