% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pathway.R
\name{add_pathway}
\alias{add_pathway}
\title{Pathway Alterations}
\usage{
add_pathway(
  bin_mat,
  pathway = c("RTK/RAS", "Nrf2", "PI3K", "TGFB", "p53", "Wnt", "Myc", "Cell cycle",
    "Hippo", "Notch", "all")
)
}
\arguments{
\item{bin_mat}{a binary matrix}

\item{pathway}{The pathways to check.  The options are "RTK/RAS", "Nrf2", "PI3K", "TGFB", "p53", "Wnt", "Myc", "Cell cycle", "Hippo", "Notch", "all".}
}
\value{
a data frame: each sample is a row, columns are pathways, with values of 0/1 depending on pathway alteration status.
}
\description{
Checks if certain oncogenic signaling pathways are altered.  Pathways were curated from \href{https://pubmed.ncbi.nlm.nih.gov/29625050/}{this paper}.  Please check for gene aliases before using.
}
\details{
Sanchez-Vega, F., Mina, M., Armenia, J., Chatila, W. K., Luna, A., La, K. C., Dimitriadoy, S., Liu, D. L., Kantheti, H. S., Saghafinia, S., Chakravarty, D., Daian, F., Gao, Q., Bailey, M. H., Liang, W. W., Foltz, S. M., Shmulevich, I., Ding, L., Heins, Z., Ochoa, A., … Schultz, N. (2018). Oncogenic Signaling Pathways in The Cancer Genome Atlas. Cell, 173(2), 321–337.e10. \url{https://doi.org/10.1016/j.cell.2018.03.035}
}
\examples{

binmat <- binary_matrix(mutation = mut, cna = cna, fusion = fusion)
pathway_df <- add_pathway(binmat, pathway = "all")

}
