% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_q_tbl_wide.R
\name{add_q_tbl_wide}
\alias{add_q_tbl_wide}
\title{Add a column of q-values to account for
multiple comparisons}
\usage{
add_q_tbl_wide(
  x,
  method = "fdr",
  pvalue_fun = NULL,
  n_comp = NULL,
  quiet = NULL
)
}
\arguments{
\item{x}{a \code{gtsummary} object}

\item{method}{String indicating method to be used for p-value
adjustment. Methods from
\link[stats:p.adjust]{stats::p.adjust} are accepted.  Default is \code{method = "fdr"}.}
}
\description{
Adjustments to p-values are performed with \link[stats:p.adjust]{stats::p.adjust}.
}
\section{Example Output}{

\if{html}{Example 1}

\if{html}{\out{
`r man_create_image_tag(file = "add_q_ex1.png", width = "65")`
}}

\if{html}{Example 2}

\if{html}{\out{
`r man_create_image_tag(file = "add_q_ex2.png", width = "60")`
}}
}

\examples{
\dontshow{if (broom.helpers::.assert_package("car", pkg_search = "gtsummary", boolean = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{
# Example 1 ----------------------------------
add_q_ex1 <-
  trial[c("trt", "age", "grade", "response")] \%>\%
  tbl_summary(by = trt) \%>\%
  add_p() \%>\%
  add_q()

# Example 2 ----------------------------------
add_q_ex2 <-
  trial[c("trt", "age", "grade", "response")] \%>\%
  tbl_uvregression(
    y = response,
    method = glm,
    method.args = list(family = binomial),
    exponentiate = TRUE
  ) \%>\%
  add_global_p() \%>\%
  add_q()
}
\dontshow{\}) # examplesIf}
}
\author{
Esther Drill, Daniel D. Sjoberg
}
\concept{tbl_regression tools}
\concept{tbl_summary tools}
\concept{tbl_svysummary tools}
\concept{tbl_uvregression tools}
