% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annotate_impact_missing}
\alias{annotate_impact_missing}
\title{IMPACT Panel Annotation of NA's}
\usage{
annotate_impact_missing(binary_matrix)
}
\arguments{
\item{binary_matrix}{a processed binary_matrix}
}
\value{
a data frame iwth NAs inserted for genes not tested for given panel versions
}
\description{
IMPACT Panel Annotation of NA's
}
