% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-matrix.R
\name{binary_matrix}
\alias{binary_matrix}
\title{Enables creation of a binary matrix from a mutation file with
a predefined list of patients (rows are patients and columns are genes)}
\usage{
binary_matrix(
  patients = NULL,
  mutation = NULL,
  mut_type = "SOMATIC",
  snp_only = FALSE,
  include_silent = FALSE,
  fusion = NULL,
  cna = NULL,
  cna_binary = TRUE,
  cna_relax = FALSE,
  specify_panel = "no",
  rm_empty = TRUE,
  recode_aliases = TRUE,
  ...
)
}
\arguments{
\item{patients}{a character vector specifying which patients should be included in the resulting data frame.
Default is NULL is which case all patients within the mutation, cna or fusions file will be used. If you specify
a vector of patients that contain patients not in any of the passed genomic data frames, 0's (or NAs when appropriate if specifying a panel) will be
returned for every column of that patient row.}

\item{mutation}{A data frame of mutations in the format of a maf file.}

\item{mut_type}{The mutation type to be used. Options are "SOMATIC", "GERMLINE" or "ALL". Note "ALL" will
keep all mutations regardless of status (not recommended). Default is SOMATIC.}

\item{snp_only}{Boolean to rather the genetics events to be kept only to be SNPs (insertions and deletions will be removed).
Default is FALSE.}

\item{include_silent}{Boolean to keep or remove all silent mutations. TRUE keeps, FALSE removes. Default is FALSE.}

\item{fusion}{A data frame of fusions. If not NULL the outcome will be added to the matrix with columns ending in ".fus".
Default is NULL.}

\item{cna}{A data frame of copy number alterations. If inputed the outcome will be added to the matrix with columns ending in ".del" and ".amp".
Default is NULL.}

\item{cna_binary}{A boolean argument specifying if the cna events should be enforced as binary. In which case separate columns for
amplifications and deletions will be created.}

\item{cna_relax}{By default this argument is set to FALSE, where only deep deletions (-2) and amplifications (2) will be annotated as events. When set to TRUE all deletions (-1 shallow and -2 deep) are counted as an event same for all gains (1 gain, 2 amplification) as an event.}

\item{specify_panel}{boolean specifying if specific IMPACT platforms should be considered. When TRUE NAs will fill the cells for genes
of patients that were not sequenced on that plaform. Default is TRUE.}

\item{rm_empty}{boolean specifying if columns with no events founds should be removed. Default is TRUE.}

\item{recode_aliases}{boolean specifying if automated gene name alias matching should be done. Default is TRUE. When TRUE
the function will check for genes that may have more than 1 name in your data using the aliases im gnomeR::impact_gene_info alias column}

\item{...}{Further arguments passed to the oncokb() function such a token}
}
\value{
a data frame with sample_id and alteration binary columns with values of 0/1
}
\description{
Enables creation of a binary matrix from a mutation file with
a predefined list of patients (rows are patients and columns are genes)
}
\examples{
mut.only <- binary_matrix(mutation = mut)

patients <- as.character(unique(mut$Tumor_Sample_Barcode))[1:200]

bin.mut <- binary_matrix(patients = patients, mutation = mut,
mut_type = "SOMATIC" ,snp_only = FALSE,
include_silent = FALSE)
bin.mut <- binary_matrix(patients = patients, mutation = mut,
mut_type = "SOMATIC", snp_only = FALSE,
include_silent = FALSE,
cna_relax = TRUE, specify_panel = "no", rm_empty = FALSE)
}
