% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-bin-mat.R
\name{binmat}
\alias{binmat}
\title{binmat}
\usage{
binmat(
  patients = NULL,
  maf = NULL,
  mut.type = "SOMATIC",
  SNP.only = F,
  include.silent = F,
  fusion = NULL,
  cna = NULL,
  cna.relax = F,
  spe.plat = F,
  set.plat = NULL,
  rm.empty = T
)
}
\arguments{
\item{patients}{a character vector that let's the user specify the patients to be used to create the matrix.
Default is NULL is which case all patients in the MAF file will be used.}

\item{maf}{A MAF file.}

\item{mut.type}{The mutation type to be used. Options are "SOMATIC", "GERMLINE" or "ALL". Note "ALL" will
keep all mutations regardless of status (not recommended). Default is SOMATIC.}

\item{SNP.only}{Boolean to rather the genetics events to be kept only to be SNPs (insertions and deletions will be removed).
Default is FALSE.}

\item{include.silent}{Boolean to keep or remove all silent mutations. TRUE keeps, FALSE removes. Default is FALSE.}

\item{fusion}{An optional MAF file for fusions. If inputed the outcome will be added to the matrix with columns ending in ".fus".
Default is NULL.}

\item{cna}{An optional CNA files. If inputed the outcome will be added to the matrix with columns ending in ".del" and ".amp".
Default is NULL. Note that this file must have patients as columns and genes as rows. binmat expects a matrix with
values between -2 and 2. Please do not use any other format. Other functions in the package are available to deal with more detailed
CNA data.}

\item{cna.relax}{for cna data only enables to count both gains and shallow deletions as amplifications and deletions respectively.}

\item{spe.plat}{boolean specifying if specific IMPACT platforms should be considered. When TRUE NAs will fill the cells for genes
of patients that were not sequenced on that plaform. Default is TRUE.}

\item{set.plat}{character argument specifying which IMPACT platform the data should be reduced to if spe.plat is set to TRUE.
Options are "341", "410" and "468". Default is NULL.}

\item{rm.empty}{boolean specifying if columns with no events founds should be removed. Default is TRUE.}
}
\value{
mut : a binary matrix of mutation data
}
\description{
Enables creation of a binary matrix from a maf file with
a predefined list of patients (rows are patients and columns are genes)
}
\examples{
library(gnomeR)
mut.only <- binmat(maf = mut)
patients <- as.character(unique(mut$Tumor_Sample_Barcode))[1:1000]
bin.mut <- binmat(patients = patients,maf = mut,mut.type = "SOMATIC",SNP.only = F,include.silent = F, spe.plat = F)
bin.mut <- binmat(patients = patients,maf = mut,mut.type = "SOMATIC",SNP.only = F,include.silent = F,
cna.relax = T, spe.plat = F, set.plat = "410", rm.empty = F)
}
