% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbp_mut}
\alias{cbp_mut}
\title{Consequence Map}
\format{
A data frame with mutations from Baca et al. Cell 2013. Retrieved from cBioPortal.
}
\usage{
cbp_mut
}
\description{
Dataset of recoding values for consequence mutation data
}
\keyword{datasets}
