% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_cna_input}
\alias{check_cna_input}
\title{Check CNA data frame to ensure columns are correct}
\usage{
check_cna_input(cna, ...)
}
\arguments{
\item{cna}{a cna data frame}

\item{...}{other arguments passed from create_gene_binary()}
}
\value{
a checked data frame
}
\description{
Check CNA data frame to ensure columns are correct
}
