% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_maf_column}
\alias{check_maf_column}
\title{Check Columns of MAF}
\usage{
check_maf_column(maf, col_to_check)
}
\arguments{
\item{maf}{Raw maf dataframe containing alteration data}

\item{col_to_check}{Which column to check}
}
\value{

}
\description{
Check Columns of MAF
}
\examples{
check_maf_column(mut, "Variant_Classification")

}
