% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_maf_input}
\alias{check_maf_input}
\title{Checks MAF input to ensure column names are correct and renamed genes are corrected}
\usage{
check_maf_input(maf)
}
\arguments{
\item{maf}{Raw maf dataframe containing alteration data}
}
\value{
a corrected maf file or an error if problems with maf
}
\description{
Checks MAF input to ensure column names are correct and renamed genes are corrected
}
\examples{

check_maf_input(mut)

}
