% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clin.patients}
\alias{clin.patients}
\title{An example of clinical patient level information file from
IMPACT cbioPortal dataset.}
\format{
A data frame with 497 observations on the following 6 variables
\describe{
\item{X.Patient.Identifier}{Factor, IMPACT patient ID}
\item{Sex}{Factor, Sex information: \code{Female}, \code{Male}}
\item{Patient.s.Vital.Status}{Factor, indicates whether the patient is alive or deceased}
\item{Smoking.History}{Factor, Indicates if person has a history of smoking}
\item{Overall.Survival..Months.}{Overall survival in months}
\item{Overall.Survival.Status}{Survival status}
}
}
\source{
\url{https://cbioportal.mskcc.org/}
}
\usage{
clin.patients
}
\description{
This set contains the publicly available data from cbioPortal patient level
clinical data. The file is organized with each line corresponding
to the information of each patient.
}
\keyword{datasets}
