% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clin.sample}
\alias{clin.sample}
\title{An example of clinical sample level information file from IMPACT cbioPortal dataset.}
\format{
A data frame with 500 observations on the following 16 variables
\describe{
\item{X.Patient.Identifier}{Factor, IMPACT patient ID}
\item{Sample.Identifier}{Factor, IMPACT sample ID}
\item{Sample.Collection.Source}{Factor, was the sample collected in-house?}
\item{Specimen.Preservation.Type}{Factor, indicates whether the patient is alive or deceased}
\item{Specimen.Type}{Factor, Indicates if person has a history of smoking}
\item{DNA.Input}{}
\item{Sample.coverage}{}
\item{Tumor.Purity}{}
\item{Matched.Status}{Overall survival in months}
\item{Sample.Type}{Survival status}
\item{Primary.Tumor.Site}{Survival status}
\item{Metastatic.Site}{Survival status}
\item{Sample.Class}{Survival status}
\item{Oncotree.Code}{Survival status}
\item{Cancer.Type}{Survival status}
\item{Cancer.Type.Detailed}{Survival status}
}
}
\source{
\url{https://cbioportal.mskcc.org/}
}
\usage{
clin.sample
}
\description{
This set contains the publicly available data from cbioPortal sample level
clinical data. The file is organized with each line corresponding
to the information of each sample.
}
\keyword{datasets}
