% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clin_collab_df}
\alias{clin_collab_df}
\title{An example data set for an IMPACT analysis coming from a clinical collaborator}
\format{
A data frame with copy number alterations (CNA) retrieved from cBioPortal.
\describe{
\item{\code{cbioportal_id}}{character with IMPACT sample ID}
\item{\code{ctype}}{character cancer type}
\item{\code{primary_mets}}{character with sample type}
}
}
\usage{
clin_collab_df
}
\description{
This set was created from a sample of 20 patients from
publicly available prostate cancer data from cBioPortal (\code{study_id = "gbc_mskcc_2022"}).
}
\keyword{datasets}
