% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cna_wide}
\alias{cna_wide}
\title{An example IMPACT cBioPortal CNA in wide format}
\format{
A data frame with copy number alterations (CNA) retrieved from cBioPortal.
}
\usage{
cna_wide
}
\description{
This set was created from a sample of 20 patients from
publicly available prostate cancer data from cBioPortal (\code{study_id = "gbc_mskcc_2022"}).
}
\keyword{datasets}
