% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consequence_map}
\alias{consequence_map}
\title{Consequence Map}
\format{
A data frame with 23878 rows of CNA data from
}
\usage{
consequence_map
}
\description{
Dataset of recoding values for consequence mutation data
}
\keyword{datasets}
