% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-bin-mat.R
\name{create.bin.matrix}
\alias{create.bin.matrix}
\title{create.bin.matrix}
\usage{
create.bin.matrix(patients = NULL, maf, mut.type = "SOMATIC",
  SNP.only = F, include.silent = F)
}
\arguments{
\item{patients}{a character vector that let's the user specify the patients to be used to create the matrix.
Default is NULL is which case all patients in the MAF file will be used.}

\item{maf}{A MAF file.}

\item{mut.type}{The mutation type to be used. Options are "SOMATIC", "GERMLINE" or "ALL". Note "ALL" will
keep all mutations regardless of status (not recommended). Default is SOMATIC.}

\item{SNP.only}{Boolean to rather the genetics events to be kept only to be SNPs (insertions and deletions will be removed).
Default is FALSE.}

\item{include.silent}{Boolean to keep or remove all silent mutations. TRUE keeps, FALSE removes. Default is FALSE.}
}
\value{
mut : a binary matrix of mutation data

no.mu.patients : a character vector of patients having no mutations found in the MAF file.
}
\description{
Enables creation of a binary matrix from a maf file with
a predifined list of patients (rows are patients and columns are genes)
}
