% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curated_genes}
\alias{curated_genes}
\title{List of curated genes for IMPACT oncoKB annotation.}
\format{
A data frame with 417 observations on the following 2 variables.
\describe{
\item{id}{}
\item{gene}{}
}
}
\usage{
curated_genes
}
\description{
We include the curated genes for the IMPACT sequencing for the oncoKB annotation pipeline.
}
\keyword{datasets}
