% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-oncoprint-dat.R
\name{dat.oncoPrint}
\alias{dat.oncoPrint}
\title{dat.oncoPrint}
\usage{
dat.oncoPrint(gen.dat, clin.dat = NULL)
}
\arguments{
\item{gen.dat}{A binary matrix or dataframe, with patients as rows and features as columns. Note that the names of the
columns must end in ".Del" or ".Amp" to recognize copy number alterations. (see create.bin.matrix for more details on this format).}

\item{clin.dat}{An optional clinical file, including only the features the user wishes to add to the plot. Default is NULL.}
}
\value{
mat : a matrix ready to be plotted using plot.Oncoprint().
}
\description{
Enables creation of a matrix used to generate an OncoPrint heatmap.
}
