% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.abort_if_not_numeric}
\alias{.abort_if_not_numeric}
\title{Check if all gene_binary columns except sample_id and other_vars are numeric}
\usage{
.abort_if_not_numeric(alt_data)
}
\arguments{
\item{alt_data}{a binary data frame created from \code{create_gene_binary()}}
}
\value{
an error message if not all columns are numeric
}
\description{
Check if all gene_binary columns except sample_id and other_vars are numeric
}
\keyword{internal}
