% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gene-binary.R
\name{.check_for_fus_in_mut}
\alias{.check_for_fus_in_mut}
\title{Check for fusions in maf file}
\usage{
.check_for_fus_in_mut(mutation)
}
\value{
a data frame if no fusions found
}
\description{
Check for fusions in maf file
}
\keyword{internal}
