% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gene-binary.R
\name{.check_for_silent}
\alias{.check_for_silent}
\title{Check for silent mutations}
\usage{
.check_for_silent(mutation, include_silent)
}
\arguments{
\item{mutation}{Raw maf dataframe containing alteration data}

\item{include_silent}{Silent mutations will be removed if FALSE (default). Variant classification column is needed.}
}
\value{
a corrected maf file or an error if problems with maf
}
\description{
Check for silent mutations
}
\keyword{internal}
