% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_gb}
\alias{.check_gb}
\title{Check that object passed is a gene binary table}
\usage{
.check_gb(data)
}
\arguments{
\item{data}{A data frame to check}
}
\value{
If data set class does not include "tbl_gene_binary" it will return an error message.
If it does, nothing will be returned.
}
\description{
Check that object passed is a gene binary table
}
\keyword{internal}
