% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_required_cols}
\alias{.check_required_cols}
\title{Check a Data Frame for Required Columns}
\usage{
.check_required_cols(data, required_cols)
}
\arguments{
\item{data}{A data frame to check}

\item{required_cols}{A character specifying names of columns to check}
}
\value{
If data set doesn't have required columns it will return an error message.
If it does have required columns, nothing will be returned
}
\description{
Check a Data Frame for Required Columns
}
\keyword{internal}
