% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-matrix.R
\name{.cna_binary_matrix}
\alias{.cna_binary_matrix}
\title{Make Binary Matrix From CNA data frame}
\usage{
.cna_binary_matrix(
  cna,
  patients = patients,
  cna_binary,
  cna_relax,
  specify_panel,
  recode_aliases
)
}
\arguments{
\item{cna}{A data frame of copy number alterations. If inputed the outcome will be added to the matrix with columns ending in ".del" and ".amp".
Default is NULL.}

\item{patients}{a character vector specifying which patients should be included in the resulting data frame.
Default is NULL is which case all patients within the mutation, cna or fusions file will be used. If you specify
a vector of patients that contain patients not in any of the passed genomic data frames, 0's (or NAs when appropriate if specifying a panel) will be
returned for every column of that patient row.}

\item{cna_binary}{A boolean argument specifying if the cna events should be enforced as binary. In which case separate columns for
amplifications and deletions will be created.}

\item{cna_relax}{By default this argument is set to FALSE, where only deep deletions (-2) and amplifications (2) will be annotated as events. When set to TRUE all deletions (-1 shallow and -2 deep) are counted as an event same for all gains (1 gain, 2 amplification) as an event.}

\item{specify_panel}{boolean specifying if specific IMPACT platforms should be considered. When TRUE NAs will fill the cells for genes
of patients that were not sequenced on that plaform. Default is TRUE.}

\item{recode_aliases}{boolean specifying if automated gene name alias matching should be done. Default is TRUE. When TRUE
the function will check for genes that may have more than 1 name in your data using the aliases im gnomeR::impact_gene_info alias column}
}
\value{
a data frame
}
\description{
Make Binary Matrix From CNA data frame
}
