% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_genomic.R
\name{.create_wide_format}
\alias{.create_wide_format}
\title{\itemize{
\item Wide format by alteration type -----
}}
\usage{
.create_wide_format(gene_binary_df, order_genes_df, table_data_df)
}
\arguments{
\item{table_data_df}{}
}
\value{
wide format 'tbl_genomic' by alteration type
}
\description{
This helper function will stratify tbl_genomic by alteration type such that
the resulting table will include one column per alteration type and an
overall summary column.
}
\keyword{internal}
