% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gene-binary.R
\name{.filter_to_sample_list}
\alias{.filter_to_sample_list}
\title{Filter to final sample list}
\usage{
.filter_to_sample_list(df, samples_final = NULL)
}
\arguments{
\item{df}{a mutation, cna, or fusion data frame to filter}

\item{samples_final}{a vector of sample IDs to filter to}
}
\value{
a filtered data frame only including select samples
}
\description{
Filter to final sample list
}
\keyword{internal}
