% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.genbin_matrix}
\alias{.genbin_matrix}
\title{Create binary matrices depending on type of mutation data}
\usage{
.genbin_matrix(data, samples, type = c("reformat_cna", "mut", "cna", "sv"))
}
\arguments{
\item{data}{a dataset of alterations}

\item{samples}{a vector of unique sample ids}

\item{type}{a character indicator for which type of alteration the dataset contains}
}
\value{
a binary matrix of alterations
}
\description{
Create binary matrices depending on type of mutation data
}
