% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gene-binary.R
\name{.infer_mutation_status}
\alias{.infer_mutation_status}
\title{Infer mutation status and assume somatic if none}
\usage{
.infer_mutation_status(mutation)
}
\arguments{
\item{mutation}{a mutation maf data frame}
}
\value{
a mutations data frame with a mutation status column
}
\description{
Infer mutation status and assume somatic if none
}
\keyword{internal}
