% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gene-binary.R
\name{.infer_variant_type}
\alias{.infer_variant_type}
\title{Infer variant type if not present in data}
\usage{
.infer_variant_type(mutation, names_mut_dict = names_mut_dict)
}
\arguments{
\item{mutation}{mutation maf file data frame}
}
\value{
a mutation data frame with a variant type column
}
\description{
Infers variant_type from reference_allele or tumor_seq_allele data
}
\keyword{internal}
