% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-matrix.R
\name{.mutations_binary_matrix}
\alias{.mutations_binary_matrix}
\title{Make Binary Matrix From Mutation data frame}
\usage{
.mutations_binary_matrix(
  mutation,
  patients = patients,
  mut_type,
  snp_only,
  include_silent,
  specify_panel,
  recode_aliases = recode_aliases
)
}
\arguments{
\item{mutation}{A data frame of mutations in the format of a maf file.}

\item{patients}{a character vector specifying which patients should be included in the resulting data frame.
Default is NULL is which case all patients within the mutation, cna or fusions file will be used. If you specify
a vector of patients that contain patients not in any of the passed genomic data frames, 0's (or NAs when appropriate if specifying a panel) will be
returned for every column of that patient row.}

\item{mut_type}{The mutation type to be used. Options are "SOMATIC", "GERMLINE" or "ALL". Note "ALL" will
keep all mutations regardless of status (not recommended). Default is SOMATIC.}

\item{snp_only}{Boolean to rather the genetics events to be kept only to be SNPs (insertions and deletions will be removed).
Default is FALSE.}

\item{include_silent}{Boolean to keep or remove all silent mutations. TRUE keeps, FALSE removes. Default is FALSE.}

\item{specify_panel}{boolean specifying if specific IMPACT platforms should be considered. When TRUE NAs will fill the cells for genes
of patients that were not sequenced on that plaform. Default is TRUE.}

\item{recode_aliases}{boolean specifying if automated gene name alias matching should be done. Default is TRUE. When TRUE
the function will check for genes that may have more than 1 name in your data using the aliases im gnomeR::impact_gene_info alias column}
}
\value{
a data frame
}
\description{
Make Binary Matrix From Mutation data frame
}
