% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-gene-binary.R
\name{.mutations_gene_binary}
\alias{.mutations_gene_binary}
\title{Make Binary Matrix From Mutation data frame}
\usage{
.mutations_gene_binary(
  mutation,
  samples,
  mut_type,
  snp_only,
  include_silent,
  specify_panel,
  recode_aliases = recode_aliases
)
}
\arguments{
\item{mutation}{A data frame of mutations in the format of a maf file.}

\item{samples}{a character vector specifying which samples should be included in the resulting data frame.
Default is NULL is which case all samples with an alteration in the mutation, cna or fusions file will be used. If you specify
a vector of samples that contain samples not in any of the passed genomic data frames, 0's (or NAs when appropriate if specifying a panel) will be
returned for every column of that patient row.}

\item{mut_type}{The mutation type to be used. Options are "omit_germline", "somatic_only", "germline_only" or "all". Note "all" will
keep all mutations regardless of status (not recommended). Default is omit_germline which includes all somatic mutations, as well as any unknown mutation types (most of which are usually somatic)}

\item{snp_only}{Boolean to rather the genetics events to be kept only to be SNPs (insertions and deletions will be removed).
Default is FALSE.}

\item{include_silent}{Boolean to keep or remove all silent mutations. TRUE keeps, FALSE removes. Default is FALSE.}

\item{specify_panel}{Default is \code{"no"} where no panel annotation is done. Otherwise pass a character vector of length 1 with a
panel id (see \code{gnomeR::gene_panels} for available panels), or \code{"impact"} for automated IMPACT annotation.
Alternatively, you may pass a data frame of \code{sample_id}-\code{panel_id} pairs specifying panels for each sample for
which to insert NAs indicating genes not tested. See below for details.}

\item{recode_aliases}{Default is \code{"impact"} where function will check for IMPACT genes that may go by more than 1 name in your data and replace the alias name with the standardized gene name (see \code{gnomeR::impact_alias_table} for reference list).
If \code{"no"}, no alias annotation will be performed.
Alternatively, you may pass a custom alias list as a data frame with columns \code{hugo_symbol} and \code{alias} specifying a custom alias table to use for checks. See below for details.}
}
\value{
a data frame
}
\description{
Make Binary Matrix From Mutation data frame
}
\keyword{internal}
