% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.recode_cna_alterations}
\alias{.recode_cna_alterations}
\title{Internal function to recode numeric CNA alteration values to factor values}
\usage{
.recode_cna_alterations(cna)
}
\arguments{
\item{cna}{a maf (long) form data set of CNAs. Must include an alteration column.}
}
\value{
a recoded CNA data set with factor alteration values
}
\description{
Internal function to recode numeric CNA alteration values to factor values
}
