% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.remove_endings}
\alias{.remove_endings}
\title{Remove descriptive endings to hugo symbol names (quicker summarize)}
\usage{
.remove_endings(names)
}
\arguments{
\item{names}{hugo symbols to check}
}
\value{
a vector of hugo symbols where each entry DOES NOT have a descriptive ending
from the following list: ".Amp", ".Del", ".fus", ".cna", ".mut".
}
\description{
Remove descriptive endings to hugo symbol names (quicker summarize)
}
\keyword{internal}
\keyword{try}
