% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize-data.R
\name{.sanitize_cna_input}
\alias{.sanitize_cna_input}
\title{Check CNA data frame to ensure columns are correct}
\usage{
.sanitize_cna_input(cna, samples_final = NULL)
}
\arguments{
\item{cna}{a cna data frame}
}
\value{
a checked data frame
}
\description{
Check CNA data frame to ensure columns are correct
}
\keyword{internal}
