% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize-data.R
\name{.sanitize_fusion_input}
\alias{.sanitize_fusion_input}
\title{Check fusion data frame to ensure columns are correct}
\usage{
.sanitize_fusion_input(fusion, samples_final = NULL)
}
\arguments{
\item{fusion}{a fusion data frame}
}
\value{
a checked data frame
}
\description{
Check fusion data frame to ensure columns are correct
}
\keyword{internal}
