% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize-data.R
\name{.sanitize_mutation_input}
\alias{.sanitize_mutation_input}
\title{Checks MAF input to ensure column names are correct and renamed genes are corrected}
\usage{
.sanitize_mutation_input(mutation, include_silent, samples_final = NULL)
}
\arguments{
\item{mutation}{Raw maf dataframe containing alteration data}

\item{include_silent}{Silent mutations will be removed if FALSE (default). Variant classification column is needed.}
}
\value{
a corrected maf file or an error if problems with maf
}
\description{
Checks MAF input to ensure column names are correct and renamed genes are corrected
}
\keyword{internal}
