% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_patient_id}
\alias{extract_patient_id}
\title{Extract IMPACT Patient ID From Sample ID}
\usage{
extract_patient_id(sample_id)
}
\arguments{
\item{sample_id}{A character vector of IMPACT Tumor sample IDs}
}
\value{
Returns a vector of patient IDs
}
\description{
Extract IMPACT Patient ID From Sample ID
}
\examples{
sample_id = c("P-0000071-T01-IM3", "P-0000072-T02-IM4", "P-0000073-T03-IM5")
extract_patient_id(sample_id)

}
