% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets.R
\name{facets_dat}
\alias{facets_dat}
\title{facets_dat}
\usage{
facets_dat(
  seg = NULL,
  filenames = NULL,
  path = NULL,
  samples = NULL,
  min_purity = 0.3,
  epsilon = 0.005,
  adaptive = FALSE
)
}
\arguments{
\item{seg}{a segmentation file containing the segmentation information of multiple samples}

\item{filenames}{the names of the segment files to be loaded and processed (Note must end in ".Rdata").}

\item{path}{the relative path to the files folder from your current directory}

\item{samples}{the names of the samples of the respective filenames. Default simply 1 to number of files.}

\item{min_purity}{the minimum purity of the sample required to be kept in the final dataset. Default is 0.3.}

\item{epsilon}{level of unions when aggregating segments between. Default is 0.005.}

\item{adaptive}{CNregions option to create adaptive segments. Default is FALSE.}
}
\value{
out.cn : a matrix of the union of all segment files with samples as rows and segments as columns

ploidy : a vector of the ploidy values for the samples in out.cn (as in facets output)

purity : a vector of the purity values for the samples in out.cn (as in facets output)

FGAs : a vector of the fragment of genome altered values for the samples in out.cn (only when tcn an lcn are available)
}
\description{
Creates a copy number alteration matrix from segment files.
}
\examples{
seg <- gnomeR::seg
samples <- unique(seg$ID)[1:100]
seg <- seg[seg$ID \%in\% samples, ]
facet <- facets_heatmap(seg = seg,
                        samples = samples)
facets_dat(seg = seg, samples = samples)
}
