% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets.R
\name{facets_heatmap}
\alias{facets_heatmap}
\title{facets_heatmap}
\usage{
facets_heatmap(
  seg = NULL,
  filenames = NULL,
  path = NULL,
  samples = NULL,
  min_purity = 0.3,
  epsilon = 0.005,
  ordered = NULL,
  outcome = NULL,
  adaptive = FALSE
)
}
\arguments{
\item{seg}{a segmentation file containing the segmentation information of multiple samples}

\item{filenames}{the names of the segment files to be loaded and processed (Note must end in ".Rdata").}

\item{path}{the relative path to the files folder from your current directory}

\item{samples}{the names of the samples of the respective filenames. Default simply 1 to number of files.}

\item{min_purity}{the minimum purity of the sample required to be kept in the final dataset. Default is 0.3.}

\item{epsilon}{level of unions when aggregating segments between}

\item{ordered}{order in which samples should be printed. Default NUll leads to hierarchical clustering.}

\item{outcome}{for seg file only, if outcome associated with study it will be printed along the x axis for each patient}

\item{adaptive}{CNregions option to create adaptive segments}
}
\value{
p a heatmap corresponding to the segment files inputted
}
\description{
Creates a heatmap of copy number profiles from segment files.
}
\examples{
library(gnomeR)
library(dplyr)
library(dtplyr)
   samples <- as.character(unique(mut$Tumor_Sample_Barcode))[1:1000]
samples.seg <- clin.sample \%>\%
  filter(Sample.Identifier \%in\% samples,
         as.numeric(as.character(Tumor.Purity)) > 30) \%>\%
  pull(Sample.Identifier)
facet <- facets_heatmap(seg = seg,
                        samples=samples.seg[0:100])
facet$p

}
