% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fusion}
\alias{fusion}
\title{An example of fusions calls file from IMPACT cbioPortal dataset}
\format{
A data frame with 127 observations with 10 variables
\describe{
 \item{Hugo_Symbol}{Factor w/ 574 levels,
   Column containing all HUGO symbols genes}
 \item{Entrez_Gene_Id}{Entrez Gene ID (NA in this dataset)}
 \item{Center}{Cancer Center ID. All are `MSKCC-DMP` in this dataset}
 \item{Tumor_Sample_Barcode}{Factor, IMPACT sample ID}
 \item{Fusion}{Factor, type of fusion}
 \item{DNA_support}{Factor, all are `yes` in this data}
 \item{RNA_support}{Factor, all are `unknown` in this data}
 \item{Method}{All are NA in this data}
 \item{Frame}{Factor, `in frame`, `out of frame` or `unknown`}
 \item{Comments}{}
}
}
\source{
\url{https://cbioportal.mskcc.org/}
}
\usage{
fusion
}
\description{
This set contains the publicly available data from cbioPortal
fusion data. The file is organized a classical MAF file.
}
\keyword{datasets}
