% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g.impact}
\alias{g.impact}
\title{List of curated genes for IMPACT}
\format{
A list of 3 items.
\describe{
 \item{g341}{}
 \item{g410}{}
 \item{g468}{}
}
}
\usage{
g.impact
}
\description{
We include the curated genes for the IMPACT sequencing from 3 different
platforms. This set of organized as a list of three different character
vectors corresponding to each of the 341, 410 and 468 gene platforms.
}
\keyword{datasets}
