% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-tab.R
\name{gen.tab}
\alias{gen.tab}
\title{gen.tab}
\usage{
gen.tab(gen.dat, outcome)
}
\arguments{
\item{gen.dat}{A matrix or dataframe, with patients as rows and features as columns.}

\item{outcome}{A leveled vector of length equal to the number of rows in gen.dat.}
}
\value{
p : an oncoprint object
}
\description{
Creates a summary table of the distribution of the genetic features by a specific outcome/covariate of interest. The difference
is tested using Fisher's exact test and further adjusted for multiple comparisons. Note that continuous genetic factors
are dichotomized at their median.
}
\examples{
library(gnomeR)
patients <- unique(clin$DMPID)[1:500]
mut.only <- create.bin.matrix(patients = patients,maf = mut)
gen.dat <- mut.only$mut
outcome = as.factor(as.character(clin$Sex[match(patients,clin$DMPID)]))
test <- gen.tab(gen.dat,outcome)
head(test)
}
