% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_panels}
\alias{gene_panels}
\title{Public Gene Panels on cBioPortal}
\format{
A nested data frame
\describe{
\item{gene_panels}{Gene panel ID}
\item{genes_in_panel}{List column of Hugo symbols of all genes in each panel}
\item{entrez_ids_in_panel}{List column of Entrez IDs of all genes in each panel}
}
}
\source{
\url{https://cbioportal.mskcc.org/}
}
\usage{
gene_panels
}
\description{
Data frame of all gene panels available in public cBioPortal
}
\keyword{datasets}
