% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genie_alias_table}
\alias{genie_alias_table}
\title{GENIE Alias Table}
\format{
A data frame with 1658 rows
\describe{
\item{hugo_symbol}{gene Hugo Symbol}
\item{alias}{Alias of Hugo Symbol in \code{hugo_symbol} column}
\item{entrez_id}{entrez ID of gene in \code{hugo_symbol}}
\item{alias_entrez_id}{entrez ID of \code{alias} gene}
}
}
\usage{
genie_alias_table
}
\description{
Data frame of genes and their aliases for all
GENIE panels. This is used for gene name resolution functionality.
}
\keyword{datasets}
