% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genie_gene_info}
\alias{genie_gene_info}
\title{GENIE consortium Gene Meta Data}
\format{
A data frame with 469 observations and 5 variables
\describe{
\item{hugo_symbol}{Column containing all HUGO symbols genes included in IMPACT}
\item{MSK_341}{Character, indicates whether each gene was included in
IMPACT platform 341. Options are \code{included} and \verb{not included}}
\item{MSK_410}{Character, indicates whether each gene was included in
IMPACT platform 410. Options are \code{included} and \verb{not included}}
\item{MSK_468}{Character, indicates whether each gene was included in
IMPACT platform 468. Options are \code{included} and \verb{not included}}
\item{DFCI_1}{Character, indicates whether each gene was included in
the original Dana Farber Cancer Institute sequencing panel. Options are \code{included} and \verb{not included}}
\item{DFCI_2}{Character, indicates whether each gene was included in
the second Dana Farber Cancer Institute sequencing panel. Options are \code{included} and \verb{not included}}
\item{DFCI_3}{Character, indicates whether each gene was included in
the third Dana Farber Cancer Institute sequencing panel. Options are \code{included} and \verb{not included}}
\item{DFCI_3.1}{Character, indicates whether each gene was included in
the updated third Dana Farber Cancer Institute sequencing panel. Options are \code{included} and \verb{not included}}
\item{UHN_48}{Character, indicates whether each gene was included in
the first 48 genes sequencing panel of the University Health Network. Options are \code{included} and \verb{not included}}
\item{UHN_50}{Character, indicates whether each gene was included in
the 50 genes sequencing panel of the University Health Network. Options are \code{included} and \verb{not included}}
\item{VICC_1_SOLIDTUMOR}{Character, indicates whether each gene was included in
the sequencing panel of the Vanderbilt-Ingram Cancer Center. Options are \code{included} and \verb{not included}}
\item{VICC_1_T7}{Character, indicates whether each gene was included in
the sequencing panel of the Vanderbilt-Ingram Cancer Center. Options are \code{included} and \verb{not included}}
\item{VICC_1_T5A}{Character, indicates whether each gene was included in
the sequencing panel of the Vanderbilt-Ingram Cancer Center. Options are \code{included} and \verb{not included}}
\item{alias}{List of EntrezIDs corresponding to the value in the hugo_symbol column and alias list.}
\item{entrez_id}{List of known aliases corresponding to the value in the hugo_symbol column.}
}
}
\source{
\url{https://cbioportal.mskcc.org/}
}
\usage{
genie_gene_info
}
\description{
Dataframe labeling all genes found in the GENIE consortium along with their corresponding
platform and Entrez ID.
}
\keyword{datasets}
