% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genie_mut}
\alias{genie_mut}
\title{An example GENIE BPC mutations data set}
\format{
A data frame with mutations retrieved using \pkg{genieBPC} package
}
\usage{
genie_mut
}
\description{
This set was created from a sample of 100 patients from the non-small cell
lung cancer v.2.0-public data set. If a row exists for a given hugo symbol
and tumor-sample-barcode, then the mutation was observed.
}
\keyword{datasets}
