% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomer_tbl_summary.R
\name{genomic_tbl_summary}
\alias{genomic_tbl_summary}
\title{genomic_tbl_summary}
\usage{
genomic_tbl_summary(data, cutoff = 0, gene_subset = NULL, by = NULL, ...)
}
\arguments{
\item{data}{data.frame of genetic samples}

\item{cutoff}{A number 0 to 1 representing the minimum percent frequency you are using to select gene's for analysis}

\item{gene_subset}{Specific genes you want to summarize, take precedent over the \code{cutoff} parameter}

\item{by}{A variable to be passed to \code{gtsummary::tbl_summary()}'s by parameter}

\item{...}{Additional parameters that can be passed to \code{gtsummary::tbl_summary()}}
}
\value{
A \code{tbl_summary()} object
}
\description{
This function will select genes based on user inputs or on frequency counts and then
will pass the data.frame to \code{gtsummary::tbl_summary()}. You can specify a \code{by} variable and other
parameters that are accepted by \code{gtsummary::tbl_summary()}. Note the \code{by} variable must be merged on to
onto the data before using the \code{by} parameter in the function.
}
\examples{
library(gnomeR)
library(gtsummary)
library(dplyr)
samples <- as.character(unique(mut$Tumor_Sample_Barcode))[1:200]
bin.mut <- binary_matrix(samples = samples, mutation = mut,
                        mut_type = "SOMATIC", snp_only = FALSE,
                        include_silent = FALSE,
                        cna_relax = TRUE, specify_panel = "no", rm_empty = FALSE)
tb1 <- genomic_tbl_summary(data = bin.mut,cutoff = .05)
tb2 <- genomic_tbl_summary(data = bin.mut,gene_subset = c("KRAS", "TERT"))
}
