% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-genetics.R
\name{get_genetics}
\alias{get_genetics}
\title{Get all genetic data for a study or list of IDs}
\usage{
get_genetics(
  sample_ids = NULL,
  sample_list_id = NULL,
  genes = "all_impact",
  database = "msk_impact",
  mutations = TRUE,
  fusions = TRUE,
  cna = TRUE
)
}
\arguments{
\item{sample_ids}{A character vector of sample ids}

\item{sample_list_id}{A character vector naming a pre-specified list of
samples (e.g. `"mskimpact_Colorectal_Cancer"`)}

\item{genes}{A list of genes to query. default is all impact genes.}

\item{database}{A character string of the database to be used. Options are "msk_impact" or "tcga", default is "msk_impact".}

\item{mutations}{Boolean specifying if mutation data should be fected. Default is TRUE.}

\item{fusions}{Boolean specifying if fusion data should be fected. Default is TRUE.}

\item{cna}{Boolean specifying if cna data should be fected. Default is TRUE.}
}
\value{
A dataframe of mutations for each sample ID
}
\description{
Get all genetic data for a study or list of IDs
}
\examples{
\dontrun{
# IMPACT #
get_genetics(sample_ids = c("P-0000004-T01-IM3", "P-0000012-T02-IM3"),
genes = 207)

# TCGA #
get_genetics(sample_ids =  c("TCGA-17-Z023-01","TCGA-02-0003-01","TCGA-02-0055-01"),
 database = "tcga")
}
}
