% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf-viz.R
\name{ggcomut}
\alias{ggcomut}
\title{Correlation Heatmap of the Top Altered Genes}
\usage{
ggcomut(mut, ...)
}
\arguments{
\item{maf}{Raw maf dataframe containing alteration data}

\item{n_genes}{Number of top genes to display in plot}
}
\value{
Correlation heatmap of the top genes
}
\description{
Correlation Heatmap of the Top Altered Genes
}
\examples{
gggenecor(mut)

}
