% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf-viz.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Heatmap of all events after binmat - using binary distance}
\usage{
ggheatmap(hmat, ...)
}
\arguments{
\item{hmat}{dataframe obtained after binamt}

\item{...}{Further arguments as passed to ComplexHeatmap::Heatmap}
}
\value{
heatmap of binmat events
}
\description{
Heatmap of all events after binmat - using binary distance
}
\examples{
set.seed(123)
patients <- as.character(unique(mut$Tumor_Sample_Barcode))[1:200]
gen.dat <- binmat(patients=patients, maf=mut, cna=cna, fusion=fusion, set.plat=TRUE)
ggheatmap(gen.dat, show_row_names=FALSE, show_column_names=FALSE)

}
