% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf-viz.R
\name{ggsnvclass}
\alias{ggsnvclass}
\title{Histogram of SNV class Counts}
\usage{
ggsnvclass(maf)
}
\arguments{
\item{maf}{Raw maf dataframe containing alteration data}
}
\value{
Histogram of counts of each SNV class
}
\description{
Histogram of SNV class Counts
}
\examples{
ggsnvclass(mut)

}
