% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{ggvarclass}
\alias{ggvarclass}
\title{Barplot of Variant Classification Counts}
\usage{
ggvarclass(mutation)
}
\arguments{
\item{mutation}{Raw mutation dataframe containing alteration data}
}
\value{
Barplot of counts of each variant classification
}
\description{
Barplot of Variant Classification Counts
}
\examples{
ggvarclass(gnomeR::mutations)

}
