% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomer_color_palette.R
\docType{data}
\name{gnomer_colors}
\alias{gnomer_colors}
\title{List of suggested color palettes for when you need a large palette}
\format{An object of class \code{character} of length 66.}
\usage{
gnomer_colors
}
\description{
Sometimes you just need a huge palette of fairly distinguishable colors.
This is a named vector of Ronglai-approved colors good for things like TCGA PanCan (33 cancer types)
or clustering solutions with high K. Run `gnomer_colors` to
see the hex codes for the study colors.
}
\keyword{datasets}
