% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomer-color-palette.R
\name{gnomer_cols}
\alias{gnomer_cols}
\title{Function to extract colors from \code{gnomer_colors} as hex codes}
\usage{
gnomer_cols(...)
}
\arguments{
\item{...}{Character names of gnomer_colors}
}
\description{
Function to extract colors from \code{gnomer_colors} as hex codes
}
