% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomer_color_palette.R
\name{gnomer_pal}
\alias{gnomer_pal}
\title{Return function to interpolate a gnomeR color palette}
\usage{
gnomer_pal(palette = "pancan", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in gnomer_palettes.
Options include "main", "pancan", "sunset"}

\item{reverse}{Boolean indicating whether the palette should be reversed.
Default to FALSE.}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a gnomeR color palette
}
