% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomer_color_palette.R
\docType{data}
\name{gnomer_palettes}
\alias{gnomer_palettes}
\title{Complete list of gnomeR color palettes}
\format{An object of class \code{list} of length 3.}
\usage{
gnomer_palettes
}
\description{
Creates color palettes based on gnomeR colors, including
"main" which is a collection of 33 pale-yet-distinct colors which flow nicely
(good for clustering solutions, for example), and
"pancan" which is a collection of 33 vibrant and distinct colors, good for visualizing the
33 TCGA PanCan cancer types.
}
\examples{
gnomer_palettes[["pancan"]]
}
\keyword{datasets}
