% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{impact_alias_table}
\alias{impact_alias_table}
\title{IMPACT Alias Tables}
\format{
A data frame with 1658 rows
\describe{
\item{hugo_symbol}{}
\item{alias}{Start }
\item{entrez_id}{End }
\item{alias_entrez_id}{}
}
}
\usage{
impact_alias_table
}
\description{
Table of aliases for IMPACT panel genes used for gene name resolution functionality.
}
\keyword{datasets}
