% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{impact_gene_info}
\alias{impact_gene_info}
\title{IMPACT Gene Meta Data}
\format{
A data frame with 574 observations and 3 variables
\describe{
  \item{hugo_symbol}{Factor w/ 574 levels,
   Column containing all HUGO symbols genes included in IMPACT}
  \item{entrez_gene_id}{Integer, contains all Entrez IDs for genes included in IMPACT}
  \item{platform_added}{Character, indicates platform at which each gene was added.
  Options are `341`, `410`, `468`}
  \item{platform_341}{Character, indicates whether each gene was included in
  IMPACT platform 341. Options are `included` and `not included`}
  \item{platform_410}{Character, indicates whether each gene was included in
  IMPACT platform 410. Options are `included` and `not included`}
  \item{platform_468}{Character, indicates whether each gene was included in
  IMPACT platform 468. Options are `included` and `not included`}
  \item{pathway}{Character, indicates whether each gene belongs to one of the pathways
  defined in Oncogenic Signaling Pathways in The Cancer Genome Atlas`}
}
}
\source{
\url{https://cbioportal.mskcc.org/}
}
\usage{
impact_gene_info
}
\description{
Dataframe labeling all genes found in IMPACT along with their corresponding
platform and Entrez ID.
}
\keyword{datasets}
