% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{impact_genes}
\alias{impact_genes}
\title{List of genes entrezID for the API}
\format{
dataframe of the IDs
\describe{
 \item{id}{EntrezID}
 \item{hugo_symbol}{Gene name}
}
}
\usage{
impact_genes
}
\description{
We include all the gene IDs in IMPACT
}
\keyword{datasets}
