% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf-summary.R
\name{maf.summary}
\alias{maf.summary}
\title{maf.summary}
\usage{
maf.summary(maf, mut.type = "SOMATIC")
}
\arguments{
\item{maf}{the names of the segment files to be loaded and processed (Note must end in ".Rdata").}

\item{mut.type}{The mutation type to be used. Options are "SOMATIC", "GERMLINE" or "ALL". Note "ALL" will
keep all mutations regardless of status (not recommended). Default is SOMATIC.}
}
\value{
A series of plot summarising information about the maf file inputted.
}
\description{
Creates a set of plot summarising a maf file.
}
\examples{
library(gnomeR)
all.plots <- maf.summary(maf=mut)
}
