% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut}
\alias{mut}
\title{An example maf file from IMPACT cbioPortal dataset}
\format{
A data frame with 457 observations with 45 variables
\describe{
}
}
\source{
\url{https://cbioportal.mskcc.org/}
}
\usage{
mut
}
\description{
This set contains the publicly available data from cbioPortal mutation data.
The file is organized a classical MAF file.
}
\keyword{datasets}
