% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutations}
\alias{mutations}
\title{An example IMPACT cBioPortal mutation data set in API format}
\format{
A data frame with mutations from Baca et al. Cell 2013. Retrieved from cBioPortal.
}
\source{
{https://pubmed.ncbi.nlm.nih.gov/32864625/}
}
\usage{
mutations
}
\description{
This set contains publicly available data from cBioPortal. The file
is in API format.
}
\keyword{datasets}
