% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_df}
\alias{names_df}
\title{Data Frame of Column Names}
\format{
A data frame with 2333 rows
\describe{
\item{maf_column_name}{MAD columns}
\item{api_column_name}{API columns}
\item{mutation_input}{Mutation columns}
\item{fusion_input}{Fusion columns}
\item{cna_input}{CNA columns}
\item{definition}{Column Definition}
\item{notes}{data notes}
\item{sc_maf_column_name}{Snake case MAF name}
\item{sc_api_column_name}{Snake case API column name}
\item{internal_column_name}{Internally used data name}
}
}
\usage{
names_df
}
\description{
Table of accepted data names
}
\keyword{datasets}
