% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf-oncokb-annotator.R
\name{oncoKB.annotate}
\alias{oncoKB.annotate}
\title{oncoKB.annotate}
\usage{
oncoKB.annotate(input, output, lib.path = NULL)
}
\value{
OncoKb annotate MAF.
}
\description{
OncoKB annotaor for MAF files.
}
