% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf-oncokb-annotator.R
\name{oncoKB_annotate}
\alias{oncoKB_annotate}
\title{oncoKB_annotate
OncoKB annotator for MAF files.}
\usage{
oncoKB_annotate(maf, cancer_types, parallelize = T)
}
\arguments{
\item{maf}{A MAF file to be annotated.}

\item{cancer_types}{Data frame with samples mapped to cancer type for accurate levels of actionability. Default is NULL.}

\item{parallelize}{Should this be done in parallel. Default is TRUE}
}
\value{
OncoKb annotate MAF.
}
\description{
oncoKB_annotate
OncoKB annotator for MAF files.
}
\examples{
library(gnomeR)
}
