% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{panel_names}
\alias{panel_names}
\title{GENIE panel names as found in synapse data}
\format{
A data frame with 469 observations and 5 variables
\describe{
\item{Sequence.Assay.ID}{Column containing all panel names as found in GENIE data}
\item{Panel}{Character, indicates corresponding names in gnomeR.}
}
}
\source{
\url{https://cbioportal.mskcc.org/}
}
\usage{
panel_names
}
\description{
Dataframe labeling all panels in the GENIE data and their corresponding names.
}
\keyword{datasets}
