% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathways}
\alias{pathways}
\title{IMPACT Oncogenic Signaling Pathways}
\format{
A list of common cancer pathways and their associated alterations
\describe{
\item{pathway}{name of pathway}
\item{genes}{vector of gene alterations in each pathways}
}
}
\source{
Sanchez-Vega, F., Mina, M., Armenia, J., Chatila, W. K., Luna, A., La, K. C., Dimitriadoy, S., Liu, D. L., Kantheti, H. S., Saghafinia, S., Chakravarty, D., Daian, F., Gao, Q., Bailey, M. H., Liang, W. W., Foltz, S. M., Shmulevich, I., Ding, L., Heins, Z., Ochoa, A., … Schultz, N. (2018). Oncogenic Signaling Pathways in The Cancer Genome Atlas. Cell, 173(2), 321–337.e10. \url{https://doi.org/10.1016/j.cell.2018.03.035}
}
\usage{
pathways
}
\description{
Oncogenic Signaling Pathways curated from \href{https://pubmed.ncbi.nlm.nih.gov/29625050/}{Sanchez-Vega, F et al., 2018}.
See \href{http://www.cbioportal.org/}{cBioPortal.org} for more information
}
\keyword{datasets}
