% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot-cna.R
\name{pivot_cna_longer}
\alias{pivot_cna_longer}
\title{Reformat Wide CNA Data to Long}
\usage{
pivot_cna_longer(wide_cna, clean_sample_ids = TRUE)
}
\arguments{
\item{wide_cna}{a cna dataframe in wide format (e.g. gnomeR::cna)}

\item{clean_sample_ids}{\code{TRUE} by default and function will clean
\code{sample_id} field to replace "." with "-". If \code{FALSE},
no modification will be made to returned \code{sample_ids} field}
}
\value{
A long data frame of CNA events
}
\description{
Takes a numeric vector of CNA data in wide format where each column is a sample
and each row is a hugo symbol. Function will return a long format CNA data set
of just events (neutral/diploid instances are filtered out) and will recode events from
numeric to descriptive (-2/-1/-1.5 is deletion, 2/1 is amplification).
}
\examples{
cna <- pivot_cna_longer(wide_cna = gnomeR::cna_wide)
}
