% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-oncoprint.R
\name{plot.oncoPrint}
\alias{plot.oncoPrint}
\title{plot.oncoPrint}
\usage{
\method{plot}{oncoPrint}(gen.dat, clin.dat = NULL, ordered = NULL)
}
\arguments{
\item{gen.dat}{A binary matrix or dataframe, with patients as rows and features as columns. Note that the names of the
columns must end in ".Del" or ".Amp" to recognize copy number alterations. (see create.bin.matrix for more details on this format).}

\item{clin.dat}{An optional clinical file, including only the features the user wishes to add to the plot. Default is NULL.}

\item{ordered}{An optional vector of length equal to the number of patients under consideration. Indicates the new order (from left to right)
to be plotted.}
}
\value{
p : an oncoprint object
}
\description{
Creates the OncoPrint corresponding to the inputted genetic data
}
