% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve-gene-aliases.R
\name{recode_alias}
\alias{recode_alias}
\title{Recode Hugo Symbol Column}
\usage{
recode_alias(genomic_df, ...)
}
\arguments{
\item{genomic_df}{a gene_binary object}

\item{...}{Other things passed}
}
\value{
A dataframe with a recoded Hugo Symbol columns
}
\description{
Searches the Hugo Symbol column in a genomic dataframe to look for
any genes that have common gene name aliases,
and replaces those aliases with the accepted (most recent) gene name.
Function uses \code{gnomeR::impact_alias_table} by default as reference for
which aliases to replace and supports IMPACT panel alias replacement only at this time.
}
\examples{
mut <- rename_columns(gnomeR::mutations)

colnames(mut)

colnames(recode_alias(genomic_df = mut))

}
