% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_cna}
\alias{recode_cna}
\title{Internal function to recode numeric CNA alteration values to factor values}
\usage{
recode_cna(alteration_vector)
}
\arguments{
\item{alteration_vector}{a vector of CNA alterations coded with any of the
following levels: neutral, deletion, amplification, gain, loss, homozygous deletion,
hemizygous deletion, loh, gain, high level amplification, 0, -1, -1.5, -2, 1, 2.}
}
\value{
a recoded CNA data set with factor alteration values. See details for code dictionary
}
\description{
Internal function to recode numeric CNA alteration values to factor values
}
\details{
CNA is coded to the following key based on key: values below
\itemize{
\item "neutral":  "0", "neutral",
\item "deletion": "homozygous deletion", "-2",
\item "deletion": "loh", "-1.5",
\item "deletion": "hemizygous deletion", "-1",
\item "amplification": "gain", "1",
\item "amplification": high level amplification", "2",
}
}
\examples{
recode_cna(gnomeR::cna$alteration[1:10])
}
