% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reformat_cna}
\alias{reformat_cna}
\title{Reformat CNA from maf version to wide version}
\usage{
reformat_cna(cna, patients = patients)
}
\arguments{
\item{cna}{a cna dataframe in maf format}

\item{patients}{a list of patients to include}
}
\value{
a dataframe of reformatted CNA alteration
}
\description{
Reformat CNA from maf version to wide version
}
