% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_fusion.R
\name{reformat_fusion}
\alias{reformat_fusion}
\title{Enables users to reformat fusions files so that each fusion is listed as one row with two hugo-symbol
sites instead of two rows, one for each site. This is the required format for the \code{create_gene_binary} function.}
\usage{
reformat_fusion(fusions)
}
\arguments{
\item{fusions}{a data frame of fusion/structural variants that occur in a cohort. There should be a \code{sample_id}, \code{hugo_symbol},
and \code{fusion} column at minimum. Intragenic/intergenic fusions will have one row. Any two gene fusions will
have two rows. See \code{gnomeR::sv_long} for an example.}
}
\value{
a data frame with \code{sample_id}, \code{site1hugo_symbol}, and \code{site2hugo_symbol} and \code{fusion} columns. This should match the format
of the \code{gnomeR::sv} dataset.
}
\description{
Enables users to reformat fusions files so that each fusion is listed as one row with two hugo-symbol
sites instead of two rows, one for each site. This is the required format for the \code{create_gene_binary} function.
}
\examples{

sv_long1 <- gnomeR::sv_long \%>\%
  rename_columns() \%>\%
  reformat_fusion()

head(sv_long1)

}
