% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomer-color-palette.R
\name{reset_gnomer_palette}
\alias{reset_gnomer_palette}
\title{Reset gnomeR color palette}
\usage{
reset_gnomer_palette(env = rlang::caller_env())
}
\arguments{
\item{env}{environment in which palette will take effect. Default is \code{rlang::caller_env()}.}
}
\description{
This function resets the gnomeR color palette back to the ggplot2 default palette for all
ggplot2 objects. A typical workflow would include this after a call to \code{set_gnomer_palette()}
function is no longer needed,
and subsequent calls to \code{ggplot()} will utilize the default color palette from ggplot2.
}
\examples{
library(ggplot2)

set_gnomer_palette()

ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) +
  geom_point()

reset_gnomer_palette()
#default reset
ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) +
  geom_point()

}
\author{
Michael Curry
}
