% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_alias}
\alias{resolve_alias}
\title{Resolve Hugo Symbol Names with Aliases}
\usage{
resolve_alias(gene_to_check, alias_table = all_alias_table)
}
\arguments{
\item{alias_table}{}
}
\value{
if the accepted hugo symbol is input, it is returned back.
If an alias name is provided, the more common name/more up to date name is returned
}
\description{
Resolve Hugo Symbol Names with Aliases
}
\examples{
resolve_alias("KMT2D", alias_table = tidyr::unnest(impact_genes, cols = alias))

}
