% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_fusion_input}
\alias{sanitize_fusion_input}
\title{Check fusion data frame to ensure columns are correct}
\usage{
sanitize_fusion_input(fusion, ...)
}
\arguments{
\item{fusion}{a fusion data frame}

\item{...}{other arguments passed from create_gene_binary()}
}
\value{
a checked data frame
}
\description{
Check fusion data frame to ensure columns are correct
}
\examples{
fus <- sanitize_fusion_input(fusion = gnomeR::fusion)

}
