% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seg}
\alias{seg}
\title{A segmentation file from the cbioPortal datasets}
\format{
A data frame with 30240 observations with 6 variables
\describe{
\item{ID}{Factor, IMPACT sample ID}
\item{chrom}{chromosome}
\item{loc.start}{start location}
\item{loc.end}{end location}
\item{num.mark}{number of probes or bins covered by the segment}
\item{seg.mean}{segment mean value, usually in log2 scale}
}
}
\source{
\url{https://cbioportal.mskcc.org/}
}
\usage{
seg
}
\description{
Segmentation file provided by the processing of IMPACT sequencing using FACETS
}
\keyword{datasets}
