% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomer-color-palette.R
\name{set_gnomer_palette}
\alias{set_gnomer_palette}
\title{Set gnomeR color palette}
\usage{
set_gnomer_palette(
  palette = c("pancan", "main", "sunset"),
  gradient = c("pancan", "main", "sunset"),
  reverse = FALSE,
  env = rlang::caller_env()
)
}
\arguments{
\item{palette}{name of palette in gnomer_palettes, supplied in quotes.
Options include \verb{"pancan", "main", "sunset"}.
Default is \code{"pancan"}.}

\item{gradient}{name of gradient palette in \code{gnomer_palettes}, supplied in quotes.
Options include \verb{"pancan", "main", "sunset"}. Default is \code{"pancan"}.}

\item{reverse}{if set to \code{TRUE}, will reverse the order of the color palette}

\item{env}{environment in which palette will take effect. Default is \code{rlang::caller_env()}.}
}
\description{
\lifecycle{experimental}
This function sets the gnomeR color palette as the default palette for all
ggplot2 objects. It does so by overriding the following four functions from
the ggplot2 package: \code{scale_color_discrete()},
\code{scale_fill_discrete()}, \code{scale_color_continuous()}, and
\code{scale_fill_continuous()}, and places them in the specified environment.
A typical workflow would include this function at the top of a script,
and subsequent calls to \code{ggplot()} will utilize the MSK color palette.
}
\examples{
library(ggplot2)

set_gnomer_palette()

ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) +
  geom_point()

# setting other MSK palettes
set_gnomer_palette(palette = "main", gradient = "sunset")

ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) +
  geom_point()

ggplot(mtcars, aes(wt, mpg, color = cyl)) +
  geom_point()
}
\author{
Michael Curry
}
