% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate-panels.R
\name{specify_impact_panels}
\alias{specify_impact_panels}
\title{IMPACT Panel Annotation of NA's}
\usage{
specify_impact_panels(gene_binary)
}
\arguments{
\item{gene_binary}{a processed gene_binary}
}
\value{
a data frame iwth NAs inserted for genes not tested for given panel versions
}
\description{
IMPACT Panel Annotation of NA's
}
