% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{substrRight}
\alias{substrRight}
\title{Utility Function to Extract SNV}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{string}

\item{n}{number of characters from right}
}
\value{
string
}
\description{
Utility Function to Extract SNV
}
\examples{
substrRight("Hello", 2)

}
