% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_by_gene.R
\name{summarize_by_gene}
\alias{summarize_by_gene}
\title{Simplify binary matrix to one column per gene that counts any alteration type as 1}
\usage{
summarize_by_gene(binary_matrix)
}
\arguments{
\item{binary_matrix}{a 0/1 matrix of gene alterations}
}
\value{
a binary matrix with a row for each sample and one column per gene
}
\description{
This will reduce the number of columns in your binary matrix, and the
resulting data frame will have only 1 col per gene, as opposed to separate
columns for mutation/cna/fusion.
}
\examples{
library(gnomeR)
samples <- as.character(unique(mut$Tumor_Sample_Barcode))[1:200]
binary_matrix <- binary_matrix(samples = samples, mutation = mut,cna = cna,
                        mut_type = "somatic_only", snp_only = FALSE,
                        include_silent = FALSE,
                        cna_relax = TRUE, specify_panel = "IMPACT341", rm_empty = FALSE) \%>\%
                 summarize_by_gene()

}
