% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sv_long}
\alias{sv_long}
\title{An example of long-format fusion/sv files}
\format{
A data frame with 30 unique \code{sample_id} values and 62 hugo symbols listed
}
\source{
\url{https://www.aacr.org/professionals/research/aacr-project-genie/bpc/}
}
\usage{
sv_long
}
\description{
This set was created from a sample of 30 patients from publicly available
non-small cell lung cancer data from GENIEBPC \verb{(cohort = 'NSCLC', version = 'v2.0-public')}
}
\keyword{datasets}
