% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_genomic.R
\name{tbl_genomic}
\alias{tbl_genomic}
\title{tbl_genomic}
\usage{
tbl_genomic(
  gene_binary,
  by = NULL,
  freq_cutoff = deprecated(),
  freq_cutoff_by_gene = deprecated(),
  gene_subset = deprecated(),
  wide_format = FALSE,
  ...
)
}
\arguments{
\item{gene_binary}{data.frame of genetic samples}

\item{by}{A variable to be passed to \code{gtsummary::tbl_summary()}'s by
parameter}

\item{freq_cutoff}{deprecated}

\item{freq_cutoff_by_gene}{deprecated}

\item{gene_subset}{deprecated}

\item{wide_format}{Specifies whether to stratify tbl_genomic by alteration
type such that the resulting table will include one column per alteration
type and an overall summary column. Default is \code{wide_format = FALSE}.}

\item{...}{Additional parameters that can be passed to
\code{gtsummary::tbl_summary()}. To access the additional parameters you need to
load \code{gtsummary}.}
}
\value{
A \code{tbl_summary()} object
}
\description{
This function will select genes based on user inputs or on frequency counts
and then will pass the data.frame to \code{gtsummary::tbl_summary()}. You can
specify a \code{by} variable and other parameters that are accepted by
\code{gtsummary::tbl_summary()}. Note the \code{by} variable must be merged on to onto
the data before using the \code{by} parameter in the function.
}
\examples{

samples <- unique(mutations$sampleId)[1:10]

gene_binary <- create_gene_binary(
  samples = samples,
  mutation = gnomeR::mutations,
  cna = gnomeR::cna,
  mut_type = "somatic_only", snp_only = FALSE,
  specify_panel = "no"
)

tbl1 <- tbl_genomic(gene_binary)

# Example wth `by` variable

gene_binary$sex <- sample( c("M", "F"), size = nrow(gene_binary), replace = TRUE)

tbl2 <- tbl_genomic(
  gene_binary = gene_binary,
  by = sex
) \%>\%
gtsummary::add_p() \%>\%
gtsummary::add_q()

}
