% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_genes}
\alias{tcga_genes}
\title{A vector of 19441 hugo symbols in TCGA}
\format{
A data frame with 19441 genes
\describe{
    }
}
\usage{
tcga_genes
}
\description{
A vector of 19441 hugo symbols in TCGA
}
\keyword{datasets}
