% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_samples}
\alias{tcga_samples}
\title{Data frame of all TCGA sample ids and their corresponding cancer type}
\format{
A data frame with 10275 samples
\describe{
\item{patient_id}{Sample ID}
\item{Cancer_Code}{Cancer site following TCGA abbreviations}
}
}
\usage{
tcga_samples
}
\description{
Data frame of all TCGA sample ids and their corresponding cancer type
}
\keyword{datasets}
