% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ti_341}
\alias{ti_341}
\title{Intervals sequenced in 341 panel}
\format{
A data frame with 6360 lines
\describe{
\item{V1}{Chromosome}
\item{V2}{Start position}
\item{V3}{End position}
\item{V4}{Strand}
\item{V5}{Interval name}
}
}
\usage{
ti_341
}
\description{
List of the genes and their corresponding positions sequenced in the targeted panel of IMPACT containing 341 genes.
}
\keyword{datasets}
