% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ti_410}
\alias{ti_410}
\title{Intervals sequenced in 410 panel}
\format{
A data frame with 7016 lines
\describe{
    \item{V1}{Chromosome}
    \item{V2}{Start position}
    \item{V3}{End position}
    \item{V4}{Strand}
    \item{V5}{Interval name}
    }
}
\usage{
ti_410
}
\description{
List of the genes and their corresponding positions sequenced in the targeted panel of IMPACT containing 410 genes.
}
\keyword{datasets}
