% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni-cox.R
\name{uni.cox}
\alias{uni.cox}
\title{uni.cox}
\usage{
uni.cox(X, surv.dat, surv.formula)
}
\arguments{
\item{X}{Matrix/surv.datframe of genomic features, continuous or binary (note cannot handle categorical surv.dat for the moment).}

\item{surv.dat}{a surv.dat frame containing the survival information. This can be made of 2 or 3 columns. 1 or 2 for time,
and one for status (where 1 is event and 0 is no event).}

\item{surv.formula}{a survival formula with names matching those in surv.dat eg: Surv(time,status)~.}
}
\value{
tab A table of all the fits performed sorted by adjusted pvalues.

p An interactive plot of log(pvalue) by hazard ration.
}
\description{
Performs univariate cox proportional hazard model on every feature
}
\examples{
library(gnomeR)
patients <- unique(clin$DMPID)[1:100]
mut.only <- create.bin.matrix(patients = patients,maf = mut)
gen.dat <- mut.only$mut
surv.dat <-clin[match(patients,clin$DMPID),match(c("time","status") ,colnames(clin))]
surv.dat$status <- ifelse(surv.dat$status == "DECEASED",1,0)
surv.dat$time <- as.numeric(as.character(surv.dat$time))

cox.fits <- uni.cox(X = gen.dat,surv.dat = surv.dat, surv.formula = Surv(time,status)~.)

}
