% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate-panels.R
\name{which_impact_panel}
\alias{which_impact_panel}
\title{provide a list of impact panels a provided gene is found within}
\usage{
which_impact_panel(hugo_symbol)
}
\arguments{
\item{hugo_symbol}{a vector of hugo symbols}
}
\value{
a data frame with hugo symbols and the IMPACT panels on which they are
included
}
\description{
provide a list of impact panels a provided gene is found within
}
\examples{

hugos <- unique(gnomeR::mutations$hugoGeneSymbol)[1:10]

which_impact_panel(hugos)

}
