% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{add.perc}
\alias{add.perc}
\title{Add a percentage to counts}
\usage{
add.perc(x, ...)
}
\arguments{
\item{x}{A barplot of the mutationviz plot that follow barplot visualization}

\item{...}{other arguments as passed to adjust the percentage label size}
}
\value{
mutationviz Barplot The same barplot is now returned with percentages
}
\description{
Add a percentage to counts
}
\examples{
ggvarclass(gnomeR::mutations) + add.perc()

}
