% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pathway.R
\name{add_pathway}
\alias{add_pathway}
\title{Pathway Alterations}
\usage{
add_pathway(
  bin_mat,
  pathway = c("RTK/RAS", "Nrf2", "PI3K", "TGFB", "p53", "Wnt", "Myc", "Cell cycle",
    "Hippo", "Notch", "all")
)
}
\arguments{
\item{bin_mat}{a binary matrix}

\item{pathway}{The pathways to check.  The options are "RTK/RAS", "Nrf2", "PI3K", "TGFB", "p53", "Wnt", "Myc", "Cell cycle", "Hippo", "Notch", "all".}
}
\value{
a data frame: each sample is a row, columns are pathways, with values of 0/1 depending on pathway alteration status.
}
\description{
Checks if certain oncogenic signaling pathways are altered.  Pathways were curated from \verb{https://pubmed.ncbi.nlm.nih.gov/29625050/}.  Please check for gene aliases before using.
}
\examples{

binmat <- binary_matrix(mutation = mut, cna = cna, fusion = fusion)
pathway_df <- add_pathway(binmat, pathway = "all")

}
