% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-pathways.R
\name{add_pathways}
\alias{add_pathways}
\title{Pathway Alterations}
\source{
Sanchez-Vega, F., Mina, M., Armenia, J., Chatila, W. K., Luna, A., La, K. C., Dimitriadoy, S., Liu, D. L., Kantheti, H. S., Saghafinia, S., Chakravarty, D., Daian, F., Gao, Q., Bailey, M. H., Liang, W. W., Foltz, S. M., Shmulevich, I., Ding, L., Heins, Z., Ochoa, A., … Schultz, N. (2018). Oncogenic Signaling Pathways in The Cancer Genome Atlas. Cell, 173(2), 321–337.e10. \url{https://doi.org/10.1016/j.cell.2018.03.035}
}
\usage{
add_pathways(
  gene_binary,
  pathways = c(names(gnomeR::pathways)),
  custom_pathways = NULL,
  count_pathways_by = c("alteration", "gene")
)
}
\arguments{
\item{gene_binary}{a binary matrix from \code{gene_binary()}}

\item{pathways}{a vector of pathway names to annotate. The options are \code{names(gnomeR::pathways)} ("RTK/RAS", "Nrf2",
"PI3K", "TGFB", "p53", "Wnt", "Myc", "Cell cycle", "Hippo", "Notch"). You can pass one pathway name, multiple pathway names, or \code{NULL}. By default, all
pathways defined in \code{gnomeR::pathways} will be included. Included default pathways are alteration-specific meaning a specific type of alteration (mut/cna/fusion)
is required to mark a 1 for that pathway. If you want gene-level pathways, use \code{summarize_by_gene()} on your binary matrix, then add
pathways you want as \code{custom_pathways}}

\item{custom_pathways}{a vector of alterations to annotate as a pathway, or a list of custom pathways (see \code{gnomeR::pathways} as example)}

\item{count_pathways_by}{Must be one of the following: "alteration" (default), or "gene". This determines whether
any type of gene alteration should be counted towards a pathway ("gene") or only specific types of alterations should be counted towards a pathway ("alteration")
By default, the function assumes alteration-specific pathway annotation and all default pathways are annotated this way. If a
custom pathway is passed with no suffix (e.g. \code{custom_pathway = 'TP53'}) it will assume it is a mutation.}
}
\value{
a data frame: each sample is a row, columns are pathways, with values of 0/1 depending on pathway alteration status.
}
\description{
Input a binary matrix of patients x genes and return a dataframe with a column per pathway
indicating if default or custom oncogenic signaling pathways
are activated in each sample. Pathways were curated
from \href{https://pubmed.ncbi.nlm.nih.gov/29625050/}{Sanchez-Vega, F et al., 2018}.
}
\details{
Please check for gene aliases in your dataset before using.
}
\examples{

gene_binary <- create_gene_binary(mutation = gnomeR::mutations,
 cna = gnomeR::cna,
 fusion = gnomeR::sv)
pathway_df <- add_pathways(gene_binary, pathways = "Notch")

}
\keyword{internal}
