% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pathways.R
\name{add_pathways}
\alias{add_pathways}
\title{Pathway Alterations}
\usage{
add_pathways(
  binary_matrix,
  pathways = c(names(gnomeR::pathways)),
  custom_pathways = NULL,
  bind_pathways = TRUE,
  count_pathways_by = c("alteration", "gene")
)
}
\arguments{
\item{binary_matrix}{a binary matrix from \code{binary_matrix()}}

\item{pathways}{a vector of pathway names to annotate. The options are \code{names(gnomeR::pathways)} ("RTK/RAS", "Nrf2",
"PI3K", "TGFB", "p53", "Wnt", "Myc", "Cell cycle", "Hippo", "Notch"). You can pass one pathway name, multiple pathway names, or \code{NULL}. By default, all
pathways defined in \code{gnomeR::pathways} will be included. Included default pathways are alteration-specific meaning a specific type of alteration (mut/cna/fusion)
is required to mark a 1 for that pathway. If you want gene-level pathways, use \code{summarize_by_gene()} on your binary matrix, then add
pathways you want as \code{custom_pathways}}

\item{custom_pathways}{a vector of alterations to annotate as a pathway, or a list of custom pathways (see \code{gnomeR::pathways} as example)}

\item{bind_pathways}{a logical indicating whether pathway columns should be joined to main binary_matrix or returned separately as a list item.
Default is TRUE and function will \code{bind_cols()} and return a data.frame. If FALSE a list will be returned.}

\item{count_pathways_by}{Must be one of the following: "alteration" (default), or "gene". This determines whether
any type of gene alteration should be counted towards a pathway ("gene") or only specific types of alterations should be counted towards a pathway ("alteration")
By default, the function assumes alteration-specific pathway annotation and all default pathways are annotated this way. If a
custom pathway is passed with no suffix (e.g. \code{custom_pathway = 'TP53'}) it will assume it is a mutation.}
}
\value{
a data frame: each sample is a row, columns are pathways, with values of 0/1 depending on pathway alteration status.
}
\description{
Checks if certain oncogenic signaling pathways are altered.  Pathways were curated from \href{https://pubmed.ncbi.nlm.nih.gov/29625050/}{this paper}.  Please check for gene aliases before using.
}
\details{
Sanchez-Vega, F., Mina, M., Armenia, J., Chatila, W. K., Luna, A., La, K. C., Dimitriadoy, S., Liu, D. L., Kantheti, H. S., Saghafinia, S., Chakravarty, D., Daian, F., Gao, Q., Bailey, M. H., Liang, W. W., Foltz, S. M., Shmulevich, I., Ding, L., Heins, Z., Ochoa, A., … Schultz, N. (2018). Oncogenic Signaling Pathways in The Cancer Genome Atlas. Cell, 173(2), 321–337.e10. \url{https://doi.org/10.1016/j.cell.2018.03.035}
}
\examples{

binary_matrix <- binary_matrix(mutation = mut, cna = cna, fusion = fusion)
pathway_df <- add_pathways(binary_matrix, pathways = "Notch")

}
