% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate-panels.R
\name{annotate_any_panel}
\alias{annotate_any_panel}
\title{Annotate Missing Gene Values According to Specific Panels}
\usage{
annotate_any_panel(sample_panel_pair, gene_binary)
}
\arguments{
\item{sample_panel_pair}{a data frame of \code{sample_id}-\code{panel_id} pairs specifying panels to use for annotation of each sample}

\item{gene_binary}{a binary matrix of 0/1 indicating alteration yes/no for each sample}
}
\value{
a gene_binary annotated for missingness
}
\description{
Annotate Missing Gene Values According to Specific Panels
}
\keyword{internal}
