% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cna}
\alias{cna}
\title{An example copy number alteration raw calls file from IMPACT cbioPortal dataset}
\format{
A data frame with 410 observations with 501 variables
\describe{
}
}
\source{
\url{https://cbioportal.mskcc.org/}
}
\usage{
cna
}
\description{
This set contains the publicly available data from cbioPortal copy number alteration
data. The file is organized with the first column being the gene name
and all other columns being samples.
}
\keyword{datasets}
