% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.genbin_matrix}
\alias{.genbin_matrix}
\title{Create empty data.frame to hold results and fill with indicators}
\usage{
.genbin_matrix(
  data,
  samples_final,
  type = c("reformat_cna", "mut", "cna", "sv")
)
}
\description{
Create empty data.frame to hold results and fill with indicators
}
