% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pathways.R
\name{.sum_alts_in_pathway}
\alias{.sum_alts_in_pathway}
\title{Title}
\usage{
.sum_alts_in_pathway(
  binary_matrix,
  pathway_list_item,
  pathway_name,
  count_pathways_by
)
}
\arguments{
\item{binary_matrix}{a binary matrix (see \code{binary_matrix()})}

\item{pathway_list_item}{a named list of length 1 with pathway name as name and vector of genes as first
and only item in list}

\item{pathway_name}{name of pathway}

\item{count_pathways_by}{passed from \code{add_pathways()}}
}
\value{
a dataframe of 1 column of 0/1s indicating pathway activated yes/no
}
\description{
Title
}
\examples{
binary_matrix <- binary_matrix(mutation = mut, cna = cna,
fusion = fusion)
x <- .sum_alts_in_pathway(binary_matrix,
 pathway_list_item = gnomeR::pathways[1],
  pathway_name = names(gnomeR::pathways[1]),
    count_pathways_by = "alteration")
}
