% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pathways.R
\name{.sum_alts_in_pathway}
\alias{.sum_alts_in_pathway}
\title{Title}
\usage{
.sum_alts_in_pathway(
  gene_binary,
  pathway_list_item,
  pathway_name,
  count_pathways_by
)
}
\arguments{
\item{gene_binary}{a binary matrix (see \code{gene_binary()})}

\item{pathway_list_item}{a named list of length 1 with pathway name as name and vector of genes as first
and only item in list}

\item{pathway_name}{name of pathway}

\item{count_pathways_by}{passed from \code{add_pathways()}}
}
\value{
a dataframe of 1 column of 0/1s indicating pathway activated yes/no
}
\description{
Title
}
\examples{
gene_binary <- create_gene_binary(mutation = gnomeR::cbp_mut, cna = gnomeR::cbp_cna,
fusion = gnomeR::cbp_sv)
x <- .sum_alts_in_pathway(gene_binary,
 pathway_list_item = gnomeR::pathways[1],
  pathway_name = names(gnomeR::pathways[1]),
    count_pathways_by = "alteration")

}
