% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genie_cna}
\alias{genie_cna}
\title{An example GENIE BPC CNA data set}
\format{
A data frame with CNA retrieved using {genieBPC} package. Column names are sample ids of cohort.
}
\usage{
genie_cna
}
\description{
This set was created from a sample of 100 patients from the non-small cell
lung cancer v.2.0-public data set.
}
\keyword{datasets}
