% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genie_fusion}
\alias{genie_fusion}
\title{An example GENIE BPC fusions data set}
\format{
A data frame with fusions retrieved using {genieBPC} package
}
\usage{
genie_fusion
}
\description{
This set was created from a sample of 100 patients from the non-small cell
lung cancer v.2.0-public data set. If a row exists for a given hugo symbol
and tumor-sample-barcode, then the mutation was observed.
}
\keyword{datasets}
