% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_tbl_summary.R
\name{genomic_tbl_summary}
\alias{genomic_tbl_summary}
\title{genomic_tbl_summary}
\usage{
genomic_tbl_summary(
  binary_matrix,
  freq_cutoff = 0,
  freq_cutoff_by_gene = TRUE,
  gene_subset = NULL,
  by = NULL,
  ...
)
}
\arguments{
\item{freq_cutoff}{A number 0 to 1 representing the minimum percent frequency you are using to select gene's for analysis}

\item{freq_cutoff_by_gene}{Logical indicating whether gene selection based on frequency \% should
be calculated at the gene level, or the alteration level. Default is TRUE, indicating all
alterations (e.g. TP53/TP53.Del/TP53.Del/TP53.fus) will be aggregated at gene level to
determine cutoff frequencies and all alteration types for top genes will be returned.}

\item{gene_subset}{Specific genes you want to summarize. This takes precedent over the \code{freq_cutoff} parameter and will
return all alterations associated with that gene (including mutations, CNA, fusions).}

\item{by}{A variable to be passed to \code{gtsummary::tbl_summary()}'s by parameter}

\item{...}{Additional parameters that can be passed to \code{gtsummary::tbl_summary()}}

\item{data}{data.frame of genetic samples}
}
\value{
A \code{tbl_summary()} object
}
\description{
This function will select genes based on user inputs or on frequency counts and then
will pass the data.frame to \code{gtsummary::tbl_summary()}. You can specify a \code{by} variable and other
parameters that are accepted by \code{gtsummary::tbl_summary()}. Note the \code{by} variable must be merged on to
onto the data before using the \code{by} parameter in the function.
}
\examples{
library(gnomeR)
library(gtsummary)
library(dplyr)
samples <- as.character(unique(mut$Tumor_Sample_Barcode))[1:200]
binary_matrix <- binary_matrix(samples = samples, mutation = mut, cna = cna,
                        mut_type = "somatic_only", snp_only = FALSE,
                        include_silent = FALSE,
                        cna_relax = TRUE, specify_panel = "no", rm_empty = FALSE)
tb1 <- genomic_tbl_summary(binary_matrix = binary_matrix, freq_cutoff = .05)
tb2 <- genomic_tbl_summary(binary_matrix = binary_matrix, gene_subset = c("KRAS", "TERT"))
}
