% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf-viz.R
\name{ggcomut}
\alias{ggcomut}
\title{Co-mutation Heatmap of the Top Altered Genes}
\usage{
ggcomut(maf, n_genes = 10, ...)
}
\arguments{
\item{maf}{Raw maf dataframe containing alteration data}

\item{n_genes}{Number of top genes to display in plot}

\item{...}{Further binmat() arguments}
}
\value{
Comutation heatmap of the top genes
}
\description{
Co-mutation Heatmap of the Top Altered Genes
}
\examples{
ggcomut(mut)

}
