% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{ggcomut}
\alias{ggcomut}
\title{Comutation Heatmap of the Top Altered Genes}
\usage{
ggcomut(mutation, n_genes = 10, ...)
}
\arguments{
\item{mutation}{Raw mutation dataframe containing alteration data}

\item{n_genes}{Number of top genes to display in plot}

\item{...}{Further create_gene_binary() arguments}
}
\value{
Comutation heatmap of the top genes
}
\description{
Comutation Heatmap of the Top Altered Genes
}
\examples{
ggcomut(mutation = gnomeR::mutations)

}
