% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gggenecor}
\alias{gggenecor}
\title{Correlation Heatmap of the Top Altered Genes}
\usage{
gggenecor(mutation, n_genes = 10, ...)
}
\arguments{
\item{mutation}{Raw mutation dataframe containing alteration data}

\item{n_genes}{Number of top genes to display in plot}

\item{...}{Further create_gene_binary() arguments}
}
\value{
Correlation heatmap of the top altered genes
}
\description{
Correlation Heatmap of the Top Altered Genes
}
\examples{
gggenecor(gnomeR::mutations)

}
