% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf-viz.R
\name{gggenecor}
\alias{gggenecor}
\title{Correlation Heatmap of the Top Altered Genes}
\usage{
gggenecor(maf, n_genes = 10, ...)
}
\arguments{
\item{maf}{Raw maf dataframe containing alteration data}

\item{n_genes}{Number of top genes to display in plot}

\item{...}{Further binary_matrix() arguments}
}
\value{
Correlation heatmap of the top altered genes
}
\description{
Correlation Heatmap of the Top Altered Genes
}
\examples{
gggenecor(mut)

}
