% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf-viz.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Heatmap of all events after binmat}
\usage{
ggheatmap(hmat, ...)
}
\arguments{
\item{hmat}{dataframe obtained after binamt}

\item{...}{Further arguments as passed to ComplexHeatmap::Heatmap}
}
\value{
heatmap of binmat events
}
\description{
Heatmap of all events after binmat
}
\examples{
set.seed(123)
patients<-as.character(unique(mut$Tumor_Sample_Barcode))[sample(1:length(unique(mut$Tumor_Sample_Barcode)), 100, replace=FALSE)]
ggheatmap(binmat(patients=patients, maf=mut, cna=cna, fusion=fusion, set.plat=TRUE), show_row_names=FALSE, show_column_names=FALSE)

}
