% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf-viz.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Heatmap of all events after binary_matrix - using binary distance}
\usage{
ggheatmap(hmat, ...)
}
\arguments{
\item{hmat}{dataframe obtained after binary_matrix()}

\item{...}{Further arguments as passed to ComplexHeatmap::Heatmap}
}
\value{
heatmap of binary_matrix events
}
\description{
Heatmap of all events after binary_matrix - using binary distance
}
\examples{
set.seed(123)
samples <- as.character(unique(mut$Tumor_Sample_Barcode))[1:200]
gen_dat <- binary_matrix(samples=samples, maf=mut, cna=cna, fusion=fusion, set.plat=TRUE)
ggheatmap(gen_dat, show_row_names=FALSE, show_column_names=FALSE)

}
