% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Heatmap of all events after gene_binary - using binary distance}
\usage{
ggheatmap(hmat, ...)
}
\arguments{
\item{hmat}{dataframe obtained after create_gene_binary()}

\item{...}{Further arguments as passed to ComplexHeatmap::Heatmap}
}
\value{
heatmap of gene_binary events
}
\description{
Heatmap of all events after gene_binary - using binary distance
}
\examples{
set.seed(123)
samples <- as.character(unique(mut$Tumor_Sample_Barcode))[1:200]
gen_dat <- create_gene_binary(samples=samples, mutation=mut, cna=cna, fusion=fusion, set.plat=TRUE)
ggheatmap(gen_dat, show_row_names=FALSE, show_column_names=FALSE)

}
