% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggsamplevar}
\alias{ggsamplevar}
\title{Histogram of Variants Per Sample Colored By Variant Classification}
\usage{
ggsamplevar(maf)
}
\arguments{
\item{maf}{Raw maf dataframe containing alteration data}
}
\value{
Histogram of counts of variants per tumor sample
}
\description{
Histogram of Variants Per Sample Colored By Variant Classification
}
\examples{
ggsamplevar(mut)

}
