% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggsnvclass}
\alias{ggsnvclass}
\title{Histogram of SNV class Counts}
\usage{
ggsnvclass(mutation)
}
\arguments{
\item{mutation}{Raw mutation dataframe containing alteration data}
}
\value{
Histogram of counts of each SNV class
}
\description{
Histogram of SNV class Counts
}
\examples{
ggsnvclass(mut)

}
