% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf-viz.R
\name{ggvartype}
\alias{ggvartype}
\title{Barplot of Variant Type Counts}
\usage{
ggvartype(maf)
}
\arguments{
\item{maf}{Raw maf dataframe containing alteration data}
}
\value{
Barplot of counts of each variant type
}
\description{
Barplot of Variant Type Counts
}
\examples{
ggvartype(mut)

}
