% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomeR-package.R
\docType{package}
\name{gnomeR-package}
\alias{gnomeR}
\alias{gnomeR-package}
\title{gnomeR: Wrangle and analyze IMPACT and TCGA mutation data}
\description{
Prepare raw mutation, cna, and fusion data for analysis. Annotate OncoKB, summarize maf files, visualize results and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MSKCC-Epi-Bio/gnomeR}
  \item \url{https://mskcc-epi-bio.github.io/gnomeR/}
  \item Report bugs at \url{https://github.com/MSKCC-Epi-Bio/gnomeR/issues}
}

}
\author{
\strong{Maintainer}: Karissa Whiting \email{karissa.whiting@gmail.com} (\href{https://orcid.org/0000-0002-4683-1868}{ORCID})

Authors:
\itemize{
  \item Jessica Lavery (\href{https://orcid.org/0000-0002-2746-5647}{ORCID})
  \item Michael Curry (\href{https://orcid.org/0000-0002-0261-4044}{ORCID})
  \item Hannah Fuchs (\href{https://orcid.org/0000-0003-2426-0096}{ORCID})
  \item Axel Martin \email{axel.steph.martin@gmail.com}
  \item Arshi Arora
}

}
\keyword{internal}
