% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomeR-package.R
\docType{package}
\name{gnomeR-package}
\alias{gnomeR}
\alias{gnomeR-package}
\title{gnomeR: Wrangle and analyze IMPACT and TCGA mutation data}
\description{
Prepare raw IMPACT and TCGA mutation data for analysis. Annotate OncoKB, summarize maf files, visualize results and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AxelitoMartin/gnomeR}
  \item Report bugs at \url{https://github.com/AxelitoMartin/gnomeR/issues}
}

}
\author{
\strong{Maintainer}: Axel Martin \email{axel.steph.martin@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Arshi Arora
  \item Karissa Whiting (\href{https://orcid.org/0000-0002-4683-1868}{ORCID})
  \item Margie Hannum (\href{https://orcid.org/0000-0002-2953-0449}{ORCID})
}

}
\keyword{internal}
