% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomeR-package.R
\docType{package}
\name{gnomeR-package}
\alias{gnomeR}
\alias{gnomeR-package}
\title{gnomeR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/margarethannum/gnomeR}
  \item Report bugs at \url{https://github.com/margarethannum/gnomeR/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
