% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomeR-package.R
\docType{package}
\name{gnomeR-package}
\alias{gnomeR}
\alias{gnomeR-package}
\title{gnomeR: Wrangle and analyze IMPACT and TCGA mutation data}
\description{
Prepare raw mutation, cna, and fusion data for analysis. Annotate OncoKB, summarize maf files, visualize results and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MSKCC-Epi-Bio/gnomeR}
  \item Report bugs at \url{https://github.com/MSKCC-Epi-Bio/gnomeR/issues}
}

}
\author{
\strong{Maintainer}: Axel Martin \email{axel.steph.martin@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Arshi Arora
}

}
\keyword{internal}
