% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{impact_alias_table}
\alias{impact_alias_table}
\title{IMPACT Alias Tables}
\format{
A data frame with 1658 rows
\describe{
\item{hugo_symbol}{gene Hugo Symbol}
\item{alias}{Alias of Hugo Symbol in \code{hugo_symbol} column}
\item{entrez_id}{entrez ID of gene in \code{hugo_symbol}}
\item{alias_entrez_id}{entrez ID of \code{alias} gene}
}
}
\usage{
impact_alias_table
}
\description{
Data frame of genes and their aliases for
IMPACT panel genes. This is used for gene name resolution functionality.
}
\keyword{datasets}
