% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{impact_gene_info}
\alias{impact_gene_info}
\title{IMPACT Gene Meta Data}
\format{
A data frame with 469 observations and 5 variables
\describe{
\item{hugo_symbol}{Column containing all HUGO symbols genes included in IMPACT}
\item{platform_341}{Character, indicates whether each gene was included in
IMPACT platform 341. Options are \code{included} and \verb{not included}}
\item{platform_410}{Character, indicates whether each gene was included in
IMPACT platform 410. Options are \code{included} and \verb{not included}}
\item{platform_468}{Character, indicates whether each gene was included in
IMPACT platform 468. Options are \code{included} and \verb{not included}}
\item{alias}{List of EntrezIDs corresponding to the value in the hugo_symbol column and alias list.}
\item{entrez_id}{List of known aliases corresponding to the value in the hugo_symbol column.}
}
}
\source{
\url{https://cbioportal.mskcc.org/}
}
\usage{
impact_gene_info
}
\description{
Dataframe labeling all genes found in IMPACT along with their corresponding
platform and Entrez ID.
}
\keyword{datasets}
