% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{impact_genes}
\alias{impact_genes}
\title{IMPACT Gene Meta Data}
\format{
A data frame of impact genes
\describe{
  \item{hugo_symbol}{Factor w/ 574 levels,
   Column containing all HUGO symbols genes included in IMPACT}
  \item{platform_341}{Character, indicates whether each gene was included in
  IMPACT platform 341. Options are `included` and `not included`}
  \item{platform_410}{Character, indicates whether each gene was included in
  IMPACT platform 410. Options are `included` and `not included`}
  \item{platform_468}{Character, indicates whether each gene was included in
  IMPACT platform 468. Options are `included` and `not included`}
  \item{alias}{List of character vectors of hugo_symbol aliases commonly used
  for IMPACT genes.}
}
}
\source{
\url{https://cbioportal.mskcc.org/}
}
\usage{
impact_genes
}
\description{
Dataframe labeling all genes found in IMPACT along with their corresponding
platform and Entrez ID.
}
\keyword{datasets}
