% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf-summary.R
\name{maf.summary}
\alias{maf.summary}
\title{maf.summary
Creates a set of plot summarising a maf file.}
\usage{
maf.summary(maf, ...)
}
\arguments{
\item{maf}{the names of the segment files to be loaded and processed (Note must end in ".Rdata").}

\item{...}{any argument belonging to the binmat method}
}
\value{
p.class Barplot of counts of each variant classification

p.type Barplot of counts of each variant type

p.SNV Histogram of counts of each SNV class

p.patient.variant Histogram of counts variants per patient

p.variant.bp Boxplot of the distribution of variant classification per patient

p.variant.dist Boxplot of the relative proportion of each variant class in individual patients

p.variant.dist.bar Stacked barplot of the variants distribution in all patients

p.SNV.dist Boxplot of the relative proportion of each SNV class in individual patients

p.corr Correlation heatmap of the top 10 genes

p.comut Heatmap of the comutation of the top 10 genes
}
\description{
maf.summary
Creates a set of plot summarising a maf file.
}
\examples{
library(gnomeR)
library(dplyr)
library(dtplyr)
patients <- as.character(unique(mut$Tumor_Sample_Barcode))[1:200]
all.plots <- maf.summary(maf=mut \%>\% filter(Tumor_Sample_Barcode \%in\% patients) )
all.plots <- maf.summary(maf=mut \%>\%
filter(Tumor_Sample_Barcode \%in\% patients),spe.plat = TRUE)
}
