% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{mutation_viz}
\alias{mutation_viz}
\title{Creates a set of plot summarising a mutation file.}
\usage{
mutation_viz(mutation, ...)
}
\arguments{
\item{mutation}{Raw mutation dataframe containing alteration data}

\item{...}{any argument belonging to the gene_binary method}
}
\value{
Returns a list of the following plots:

varclass Barplot of counts of each variant classification

vartype Barplot of counts of each variant type

snvclass Histogram of counts of each SNV class

samplevar Histogram of counts variants per patient

topgenes Barplot of counts of top variant genes

genecor Correlation heatmap of the top 10 genes
}
\description{
Creates a set of plot summarising a mutation file.
}
\examples{
mutation_viz(gnomeR::mutations)

}
