% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutations}
\alias{mutations}
\title{An example IMPACT cBioPortal mutation data set in API format}
\format{
A data frame with mutations from Abida et al. JCO Precis Oncol 2017.
Retrieved from cBioPortal.There are 725 observations and 29 variables.
\describe{
\item{hugoGeneSymbol}{Character w/ 324 levels,
Column containing all HUGO symbols genes}
\item{entrezGeneId}{Entrez Gene ID}
\item{sampleId}{MSKCC Sample ID}
\item{patientId}{Patient ID}
\item{studyId}{Indicator for Abida et al. 2017 study}
\item{center}{Cancer Center ID}
\item{mutationStatus}{Somatic or germ-line mutation status}
\item{variantType}{Mutation variant type}
\item{chr}{Chromosome mutation observed on}
\item{endPosition}{End Position}
\item{fisValue}{fisValue}
\item{functionalImpactScore}{functionalImpactScore}
\item{keyword}{keyword}
\item{linkMsa}{linkMsa}
\item{linkPdb}{linkPdb}
\item{linkXvar}{linkXvar}
\item{molecularProfileId}{molecularProfileId}
\item{mutationType}{mutationType}
\item{ncbiBuild}{ncbiBuild}
\item{proteinChange}{proteinChange}
\item{proteinPosEnd}{proteinPosEnd}
\item{proteinPosStart}{proteinPosStart}
\item{referenceAllele}{referenceAllele}
\item{refseqMrnaId}{refseqMrnaId}
\item{startPosition}{startPosition}
\item{uniquePatientKey}{uniquePatientKey}
\item{uniqueSampleKey}{uniqueSampleKey}
\item{validationStatus}{validationStatus}
\item{variantAllele}{variantAllele}}
}
\source{
\url{https://www.cbioportal.org/study/summary?id=prad_mskcc_2017}
}
\usage{
mutations
}
\description{
This set contains a random sample of 200 patients from
publicly available prostate cancer data from cBioPortal. The file
is in API format.
}
\keyword{datasets}
