% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf-oncokb-annotator.R
\name{oncoKB.annotate}
\alias{oncoKB.annotate}
\title{oncoKB.annotate}
\usage{
oncoKB.annotate(input, output, lib.path = NULL)
}
\arguments{
\item{maf}{**NEED DESCRIPTION**}
}
\value{
OncoKb annotate MAF.
}
\description{
OncoKB annotator for MAF files.
}
\examples{

# NEED EXAMPLE, doesn't necessarily have to run
}
