% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf-oncokb-annotator.R
\name{oncoKB.annotate}
\alias{oncoKB.annotate}
\title{oncoKB.annotate
OncoKB annotator for MAF files.}
\usage{
oncoKB.annotate(input, output, lib.path = NULL)
}
\arguments{
\item{input}{A MAF file to be annotated.}

\item{output}{Name of the output file}

\item{lib.path}{Path to the}
}
\value{
OncoKb annotate MAF.
}
\description{
oncoKB.annotate
OncoKB annotator for MAF files.
}
\examples{
# Needs to be run on LINUX with python
# oncoKB.annotate(input = "ExampleMaf", out = "ExampleOut")
}
