% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathways}
\alias{pathways}
\title{IMPACT Gene Pathways}
\format{
A data frame of impact genes
\describe{
  \item{pathway}{name of pathway}
  \item{genes}{vectors of genes in each pathways}
}
}
\source{
\url{https://cbioportal.mskcc.org/}
}
\usage{
pathways
}
\description{
Dataframe labeling all genes found in IMPACT along with their corresponding
platform and Entrez ID.
}
\keyword{datasets}
