% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_cna.R
\name{pivot_cna_longer}
\alias{pivot_cna_longer}
\title{Reformat Wide CNA Data to Long}
\usage{
pivot_cna_longer(wide_cna, clean_sample_ids = TRUE)
}
\arguments{
\item{wide_cna}{a cna dataframe in wide format (e.g. gnomeR::cna)}

\item{clean_sample_ids}{\code{TRUE} by default and function will clean
\code{sample_id} field to replace "." with "-". If \code{FALSE},
no modification will be made to returned \code{sample_ids} field}
}
\value{
A long data frame of CNA events
}
\description{
Reformat Wide CNA Data to Long
}
\examples{
\dontrun{
# NEEDS TO BE UPDATED
cna <- pivot_cna_longer(wide_cna = gnomeR::cna)
}
}
