% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprint.R
\name{plot_oncoprint}
\alias{plot_oncoprint}
\title{Creates the OncoPrint corresponding to the inputted genetic data}
\usage{
plot_oncoprint(
  gen_dat,
  clin_dat = NULL,
  ordered_samples = NULL,
  rank_genes = TRUE,
  background_color = "#CCCCCC"
)
}
\arguments{
\item{gen_dat}{A binary matrix or dataframe, with samples as rows and features as columns. Note that the names of the
columns must end in ".Del" or ".Amp" to recognize copy number alterations. (see create.bin.matrix for more details on this format).}

\item{clin_dat}{An optional clinical file, including only the features the user wishes to add to the plot. Default is NULL.}

\item{ordered_samples}{An optional vector of length equal to the number of samples under consideration. Indicates the new order (from left to right)
to be plotted.}

\item{rank_genes}{Boolean argument specifying if the genes should be ranked by event frequency. Default is TRUE.}

\item{background_color}{Color name to be used to fill the background of the OncoPrint.}
}
\value{
an oncoprint object
}
\description{
Creates the OncoPrint corresponding to the inputted genetic data
}
\examples{
library(gnomeR)
}
