% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve-gene-aliases.R
\name{recode_alias}
\alias{recode_alias}
\title{Recode Hugo Symbol Column}
\usage{
recode_alias(genomic_df, default = T, custom_table, ...)
}
\arguments{
\item{genomic_df}{a gene_binary object}

\item{default}{logical value if default IMPACT aliases should be used}

\item{custom_table}{a dataframe with at least two columns (hugo_symbol,
alias) with one row for each pair (only
include a \code{custom_table} if \code{default = F})}

\item{...}{Other things passed}
}
\value{
A dataframe with recoded Hugo Symbol columns
}
\description{
Searches the Hugo Symbol column in a genomic dataframe to look for
any genes that have common gene name aliases,
and replaces those aliases with the accepted (most recent) gene name.
Function uses \code{gnomeR::impact_alias_table} by default as reference for
which aliases to replace and supports IMPACT panel alias replacement only at this time.
Custom tables can be provided as long as \code{hugo_symbol} and \code{alias} columns exist.
}
