% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_alias}
\alias{recode_alias}
\title{Recode Hugo Symbol Column}
\usage{
recode_alias(genomic_df, ...)
}
\arguments{
\item{genomic_df}{a gene_binary object}

\item{...}{Other things passed}
}
\value{
A dataframe with a recoded Hugo Symbol columns
}
\description{
Recode Hugo Symbol Column
}
\examples{
mut <- rename_columns(gnomeR::mutations)

colnames(mut)

colnames(recode_alias(genomic_df = mut))

}
