% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve-gene-aliases.R
\name{recode_alias}
\alias{recode_alias}
\title{Recode Hugo Symbol Column}
\usage{
recode_alias(genomic_df, alias_table = "impact", supress_warnings = FALSE)
}
\arguments{
\item{genomic_df}{a gene_binary object}

\item{alias_table}{a string indicating "impact" or "genie", or a dataframe with at least two columns (\code{hugo_symbol},
\code{alias}) with one row for each pair.}

\item{supress_warnings}{If TRUE, function will return a list containing a dataframe of recoded results and a names vector of recoded aliases in data}
}
\value{
A dataframe with recoded Hugo Symbol columns
}
\description{
Searches the Hugo Symbol column in a genomic dataframe to look for
any genes that have common gene name aliases,
and replaces those aliases with the accepted (most recent) gene name.
Function uses \code{gnomeR::impact_alias_table} by default as reference for
which aliases to replace and supports IMPACT panel alias replacement only at this time.
Custom tables can be provided as long as \code{hugo_symbol} and \code{alias} columns exist.
}
\examples{
genomic_df <- rename_columns(gnomeR::mutations[1:5, ])

alias_table <- data.frame("hugo_symbol" = c("New Symbol", "New Symbol2"),
"alias" = c("PARP1", "AKT1"))

recode_alias(genomic_df, alias_table)
recode_alias(genomic_df, alias_table, supress_warnings = TRUE)
}
