% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_cna.R
\name{reformat_cna}
\alias{reformat_cna}
\title{Reformat CNA from maf (long) version to wide version}
\usage{
reformat_cna(cna)
}
\arguments{
\item{cna}{a cna dataframe in maf (long) format}
}
\value{
a dataframe of reformatted CNA alteration (in wide format)
}
\description{
Reformat CNA from maf (long) version to wide version
}
\examples{
   cna_long <- data.frame(
    sampleId = c("P-0001276-T01-IM3","P-0001276-T01-IM3",
                 "P-0005436-T01-IM3",
                 "P-0001276-T01-IM3","P-0003333-T01-IM3"),
    Hugo_Symbol = c("MLL2","KMT2D","HIST1H2BD",
                    "HIST1H3B","KDR"),
    alteration = c("AMPLIFICATION","AMPLIFICATION",
                   "AMPLIFICATION","AMPLIFICATION","DELETION"))

cna <- reformat_cna(cna_long)


}
