% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename columns from API results to work with gnomeR functions}
\usage{
rename_columns(df_to_check)
}
\arguments{
\item{df_to_check}{A data frame to check and recode names as needed}
}
\value{
a renamed data frame
}
\description{
Will return a named vector of internal column names as values and original data set names
as names as an attribute (\code{attr(x, "names_dict")})
}
\examples{

rename_columns(df_to_check = gnomeR::mutations)
x <- rename_columns(df_to_check = gnomeR::sv)
attr(x, "names_dict")
}
