% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename columns from API results to work with gnomeR functions}
\usage{
rename_columns(df_to_check)
}
\arguments{
\item{mutations}{a mutations data frame from {cbioportalR}}
}
\value{
a renamed data frame
}
\description{
Rename columns from API results to work with gnomeR functions
}
