% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_alias}
\alias{resolve_alias}
\title{Resolve Hugo Symbol Names with Aliases}
\usage{
resolve_alias(gene_to_check, alias_table)
}
\arguments{
\item{gene_to_check}{hugo_symbol to be check}

\item{alias_table}{table containing all the aliases}
}
\value{
if the accepted hugo symbol is input, it is returned back.
If an alias name is provided, the more common name/more up to date name is returned
}
\description{
Resolve Hugo Symbol Names with Aliases
}
\examples{
resolve_alias("MLL4", alias_table = impact_alias_table)

}
