% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{specify_impact_panels}
\alias{specify_impact_panels}
\title{IMPACT Panel Annotation of NA's}
\usage{
specify_impact_panels(binary_matrix)
}
\arguments{
\item{binary_matrix}{a processed binary_matrix}
}
\value{
a data frame iwth NAs inserted for genes not tested for given panel versions
}
\description{
IMPACT Panel Annotation of NA's
}
