% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-by-frequency.R
\name{subset_by_frequency}
\alias{subset_by_frequency}
\title{Subset a Binary Matrix By Alteration Frequency Threshold}
\usage{
subset_by_frequency(gene_binary, t = 0.1)
}
\arguments{
\item{gene_binary}{A data frame with a row for each sample and column for each
alteration. Data frame must have a \code{sample_id} column and columns for each alteration
with values of 0, 1 or NA.}

\item{t}{Threshold value between 0 and 1 to subset by. Default is 10\% (.1).}
}
\value{
a data frame with a \code{sample_id} column and columns for
alterations over the given prevalence threshold of \code{t}.
}
\description{
Subset a Binary Matrix By Alteration Frequency Threshold
}
\examples{
samples <- unique(gnomeR::mutations$sampleId)
 gene_binary <- create_gene_binary(
   samples = samples, mutation = mutations, cna = cna,
   mut_type = "somatic_only",
   include_silent = FALSE,
   specify_panel = "impact"
 )
gene_binary \%>\%
 subset_by_frequency()

}
