% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_genomic.R
\name{tbl_genomic}
\alias{tbl_genomic}
\title{tbl_genomic}
\usage{
tbl_genomic(
  gene_binary,
  freq_cutoff = 0,
  freq_cutoff_by_gene = TRUE,
  gene_subset = NULL,
  by = NULL,
  ...
)
}
\arguments{
\item{gene_binary}{data.frame of genetic samples}

\item{freq_cutoff}{A number 0 to 1 representing the minimum percent frequency you are using to select gene's for analysis.
Frequencies can be calculated at gene level, or alteration level (see \code{freq_cutoff_by_gene}).}

\item{freq_cutoff_by_gene}{Logical indicating whether gene selection based on frequency \% should
be calculated at the gene level, or the alteration level. Default is TRUE, indicating all
alterations (e.g. TP53/TP53.Del/TP53.Del/TP53.fus) will be aggregated at gene level to
determine cutoff frequencies and all alteration types for top genes will be returned.}

\item{gene_subset}{Specific genes you want to summarize. This takes precedent over the \code{freq_cutoff} parameter and will
return all alterations associated with that gene (including mutations, CNA, fusions).}

\item{by}{A variable to be passed to \code{gtsummary::tbl_summary()}'s by parameter}

\item{...}{Additional parameters that can be passed to \code{gtsummary::tbl_summary()}}
}
\value{
A \code{tbl_summary()} object
}
\description{
This function will select genes based on user inputs or on frequency counts and then
will pass the data.frame to \code{gtsummary::tbl_summary()}. You can specify a \code{by} variable and other
parameters that are accepted by \code{gtsummary::tbl_summary()}. Note the \code{by} variable must be merged on to
onto the data before using the \code{by} parameter in the function.
}
\examples{
library(dplyr)
samples <- unique(mutations$sample_id)[1:10]

gene_binary <- create_gene_binary(
  samples = samples,
  mutation = gnomeR::mutations,
  cna = gnomeR::cna,
  mut_type = "somatic_only", snp_only = FALSE,
  specify_panel = "no"
)
tb1 <- tbl_genomic(gene_binary = gene_binary, freq_cutoff = .05)
tb2 <- tbl_genomic(gene_binary = gene_binary, gene_subset = c("BRCA1", "FGFR2"))

gene_binary <- gene_binary \%>\%
  mutate(sex = sample(
    x = c("M", "F"),
    size = nrow(gene_binary), replace = TRUE
  ))

t1 <- tbl_genomic(
  gene_binary = gene_binary,
  by = sex,
  freq_cutoff = .2,
  freq_cutoff_by_gene = FALSE
)

}
